Fix conflicts after merge with e2e_refactor Signed-off-by: Trayan Manolov <trayan.manolov@limechain.tech>
