+++ e2e/e2e_test.go
@@ -20,46 +20,32 @@ import (
 	"context"
 	"database/sql"
 	"encoding/base64"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"math"
 	"math/big"
-	"reflect"
 	"strconv"
-	"strings"
 	"testing"
 	"time"
 
-	read_only "github.com/limechain/hedera-eth-bridge-validator/app/services/read-only"
+	evmSetup "github.com/limechain/hedera-eth-bridge-validator/e2e/setup/evm"
 
-	"github.com/ethereum/go-ethereum/accounts/abi"
-	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/limechain/hedera-eth-bridge-validator/app/clients/evm"
-	"github.com/limechain/hedera-eth-bridge-validator/app/clients/evm/contracts/router"
-	"github.com/limechain/hedera-eth-bridge-validator/app/clients/evm/contracts/werc721"
-	"github.com/limechain/hedera-eth-bridge-validator/app/clients/evm/contracts/wtoken"
-	mirror_node "github.com/limechain/hedera-eth-bridge-validator/app/clients/hedera/mirror-node"
 	"github.com/limechain/hedera-eth-bridge-validator/app/clients/hedera/mirror-node/model/transaction"
 	hederahelper "github.com/limechain/hedera-eth-bridge-validator/app/helper/hedera"
 	"github.com/limechain/hedera-eth-bridge-validator/app/helper/timestamp"
 	auth_message "github.com/limechain/hedera-eth-bridge-validator/app/model/auth-message"
+	"github.com/limechain/hedera-eth-bridge-validator/app/persistence/entity"
 	"github.com/limechain/hedera-eth-bridge-validator/app/persistence/entity/schedule"
 	"github.com/limechain/hedera-eth-bridge-validator/app/persistence/entity/status"
 	"github.com/limechain/hedera-eth-bridge-validator/constants"
-	"github.com/limechain/hedera-eth-bridge-validator/e2e/util"
-
-	"github.com/limechain/hedera-eth-bridge-validator/app/domain/service"
-	"github.com/limechain/hedera-eth-bridge-validator/app/persistence/entity"
-	"github.com/limechain/hedera-eth-bridge-validator/e2e/service/database"
+	"github.com/limechain/hedera-eth-bridge-validator/e2e/helper/compat"
+	"github.com/limechain/hedera-eth-bridge-validator/e2e/helper/expected"
+	"github.com/limechain/hedera-eth-bridge-validator/e2e/helper/fetch"
+	"github.com/limechain/hedera-eth-bridge-validator/e2e/helper/submit"
+	"github.com/limechain/hedera-eth-bridge-validator/e2e/helper/verify"
 	"github.com/limechain/hedera-eth-bridge-validator/e2e/setup"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/hashgraph/hedera-sdk-go/v2"
-	validatorproto "github.com/limechain/hedera-eth-bridge-validator/proto"
-	"google.golang.org/protobuf/proto"
 )
 
 var (
@@ -70,54 +56,66 @@ var (
 func Test_HBAR(t *testing.T) {
 	amount := int64(1000000000) // 10 HBAR
 	setupEnv := setup.Load()
-	now = time.Now()
+	now := time.Now()
 
 	chainId := setupEnv.Scenario.FirstEvmChainId
 	evm := setupEnv.Clients.EVM[chainId]
 	receiver := evm.Receiver
 	memo := fmt.Sprintf("%d-%s", chainId, evm.Receiver.String())
 
-	targetAsset, err := setup.NativeToWrappedAsset(setupEnv.AssetMappings, constants.HederaNetworkId, chainId, constants.Hbar)
+	targetAsset, err := evmSetup.NativeToWrappedAsset(setupEnv.AssetMappings, constants.HederaNetworkId, chainId, constants.Hbar)
 	if err != nil {
 		t.Fatalf("Expecting Token [%s] is not supported. - Error: [%s]", constants.Hbar, err)
 	}
 
-	mintAmount, fee := calculateReceiverAndFeeAmounts(setupEnv, constants.Hbar, amount)
+	mintAmount, fee := expected.ReceiverAndFeeAmounts(setupEnv.Clients.FeeCalculator, setupEnv.Clients.Distributor, constants.Hbar, amount)
 
 	// Step 1 - Verify the transfer of Hbars to the Bridge Account
-	transactionResponse, wrappedBalanceBefore := verifyTransferToBridgeAccount(setupEnv, targetAsset, evm, memo, receiver, amount, t)
+	transactionResponse, wrappedBalanceBefore := verify.TransferToBridgeAccount(t, setupEnv.Clients.Hedera, setupEnv.BridgeAccount, targetAsset, evm, memo, receiver, amount)
 
 	// Step 2 - Verify the submitted topic messages
-	receivedSignatures := verifyTopicMessages(setupEnv, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), t)
+	receivedSignatures := verify.TopicMessages(t, setupEnv.Clients.Hedera, setupEnv.TopicID, setupEnv.Scenario.ExpectedValidatorsCount, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String())
 
 	// Step 3 - Validate fee scheduled transaction
-	scheduledTxID, scheduleID := validateMembersScheduledTxs(setupEnv, constants.Hbar, generateMirrorNodeExpectedTransfersForHederaTransfer(setupEnv, constants.Hbar, fee), t)
+	expectedTransfers := expected.MirrorNodeExpectedTransfersForHederaTransfer(setupEnv.Members, setupEnv.BridgeAccount, constants.Hbar, fee)
+	scheduledTxID, scheduleID := verify.MembersScheduledTxs(t, setupEnv.Clients.Hedera, setupEnv.Clients.MirrorNode, setupEnv.Members, constants.Hbar, expectedTransfers, now)
 
 	// Step 4 - Verify Transfer retrieved from Validator API
-	transactionData := verifyFungibleTransferFromValidatorAPI(setupEnv, evm, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), constants.Hbar, fmt.Sprint(mintAmount), targetAsset, t)
+	transactionData := verify.FungibleTransferFromValidatorAPI(
+		t,
+		setupEnv.Clients.ValidatorClient,
+		setupEnv.TokenID,
+		evm,
+		hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(),
+		constants.Hbar,
+		fmt.Sprint(mintAmount),
+		targetAsset,
+	)
 
 	// Step 4.1 - Get the consensus timestamp of transfer tx
 	tx, err := setupEnv.Clients.MirrorNode.GetSuccessfulTransaction(hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String())
 	if err != nil {
 		t.Fatal("failed to get successful transaction", err)
 	}
+
 	nanos, err := timestamp.FromString(tx.ConsensusTimestamp)
 	if err != nil {
 		t.Fatal("failed to parse consensus timestamp", err)
 	}
+
 	ts := timestamp.FromNanos(nanos)
 
 	// Step 5 - Submit Mint transaction
-	txHash := submitMintTransaction(evm, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), transactionData, common.HexToAddress(targetAsset), t)
+	txHash := submit.MintTransaction(t, evm, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), transactionData, common.HexToAddress(targetAsset))
 
 	// Step 6 - Wait for transaction to be mined
-	waitForTransaction(evm, txHash, t)
+	submit.WaitForTransaction(t, evm, txHash)
 
 	// Step 7 - Validate Token balances
-	verifyWrappedAssetBalance(evm, targetAsset, big.NewInt(mintAmount), wrappedBalanceBefore, receiver, t)
+	verify.WrappedAssetBalance(t, evm, targetAsset, big.NewInt(mintAmount), wrappedBalanceBefore, receiver)
 
 	// Step 8 - Prepare Comparable Expected Transfer Record
-	expectedTxRecord := util.PrepareExpectedTransfer(
+	expectedTxRecord := expected.FungibleTransferRecord(
 		constants.HederaNetworkId,
 		chainId,
 		constants.HederaNetworkId,
@@ -129,13 +127,16 @@ func Test_HBAR(t *testing.T) {
 		receiver.String(),
 		status.Completed,
 		setupEnv.Clients.Hedera.GetOperatorAccountID().String(),
-		entity.NanoTime{Time: ts})
+		entity.NanoTime{Time: ts},
+	)
+
 	// and:
-	expectedFeeRecord := util.PrepareExpectedFeeRecord(
+	expectedFeeRecord := expected.FeeRecord(
 		scheduledTxID,
 		scheduleID,
 		fee,
-		hederahelper.ToMirrorNodeTransactionID(transactionResponse.TransactionID.String()))
+		hederahelper.ToMirrorNodeTransactionID(transactionResponse.TransactionID.String()),
+	)
 
 	authMsgBytes, err := auth_message.EncodeFungibleBytesFrom(
 		expectedTxRecord.SourceChainID,
@@ -150,64 +151,76 @@ func Test_HBAR(t *testing.T) {
 		t.Fatalf("[%s] - Failed to encode the authorisation signature. Error: [%s]", expectedTxRecord.TransactionID, err)
 	}
 
-	time.Sleep(20 * time.Second)
+	time.Sleep(30 * time.Second)
 
 	// Step 9 - Verify Database Records
-	verifyTransferRecordAndSignatures(setupEnv.DbValidator, expectedTxRecord, authMsgBytes, receivedSignatures, t)
+	verify.TransferRecordAndSignatures(t, setupEnv.DbValidator, expectedTxRecord, authMsgBytes, receivedSignatures)
 	// and:
-	verifyFeeRecord(setupEnv.DbValidator, expectedFeeRecord, t)
+	verify.FeeRecord(t, setupEnv.DbValidator, expectedFeeRecord)
 }
 
 // Test_E2E_Token_Transfer recreates a real life situation of a user who wants to bridge a Hedera native token to the EVM Network infrastructure. The wrapped token on the EVM network(corresponding to the native Hedera Hashgraph's one) gets minted, then transferred to the recipient account on the EVM network.
 func Test_E2E_Token_Transfer(t *testing.T) {
 	amount := int64(1000000000) // 10 HBAR
 	setupEnv := setup.Load()
-	now = time.Now()
+	now := time.Now()
 
 	chainId := setupEnv.Scenario.FirstEvmChainId
 	evm := setupEnv.Clients.EVM[chainId]
 	memo := fmt.Sprintf("%d-%s", chainId, evm.Receiver.String())
-	mintAmount, fee := calculateReceiverAndFeeAmounts(setupEnv, setupEnv.TokenID.String(), amount)
+	mintAmount, fee := expected.ReceiverAndFeeAmounts(setupEnv.Clients.FeeCalculator, setupEnv.Clients.Distributor, setupEnv.TokenID.String(), amount)
 
-	targetAsset, err := setup.NativeToWrappedAsset(setupEnv.AssetMappings, constants.HederaNetworkId, chainId, setupEnv.TokenID.String())
+	targetAsset, err := evmSetup.NativeToWrappedAsset(setupEnv.AssetMappings, constants.HederaNetworkId, chainId, setupEnv.TokenID.String())
 	if err != nil {
 		t.Fatalf("Expecting Token [%s] is not supported. - Error: [%s]", constants.Hbar, err)
 	}
 
 	// Step 1 - Verify the transfer of HTS to the Bridge Account
-	transactionResponse, wrappedBalanceBefore := verifyTokenTransferToBridgeAccount(setupEnv, targetAsset, setupEnv.TokenID, evm, memo, evm.Receiver, amount, t)
+	transactionResponse, wrappedBalanceBefore := verify.TokenTransferToBridgeAccount(t, setupEnv.Clients.Hedera, setupEnv.BridgeAccount, targetAsset, setupEnv.TokenID, evm, memo, evm.Receiver, amount)
 
 	// Step 2 - Verify the submitted topic messages
-	receivedSignatures := verifyTopicMessages(setupEnv, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), t)
+	receivedSignatures := verify.TopicMessages(t, setupEnv.Clients.Hedera, setupEnv.TopicID, setupEnv.Scenario.ExpectedValidatorsCount, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String())
 
 	// Step 3 - Validate fee scheduled transaction
-	scheduledTxID, scheduleID := validateMembersScheduledTxs(setupEnv, setupEnv.TokenID.String(), generateMirrorNodeExpectedTransfersForHederaTransfer(setupEnv, setupEnv.TokenID.String(), fee), t)
+	expectedTransfers := expected.MirrorNodeExpectedTransfersForHederaTransfer(setupEnv.Members, setupEnv.BridgeAccount, setupEnv.TokenID.String(), fee)
+	scheduledTxID, scheduleID := verify.MembersScheduledTxs(t, setupEnv.Clients.Hedera, setupEnv.Clients.MirrorNode, setupEnv.Members, setupEnv.TokenID.String(), expectedTransfers, now)
 
 	// Step 4 - Verify Transfer retrieved from Validator API
-	transactionData := verifyFungibleTransferFromValidatorAPI(setupEnv, evm, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), setupEnv.TokenID.String(), fmt.Sprint(mintAmount), targetAsset, t)
+	transactionData := verify.FungibleTransferFromValidatorAPI(
+		t,
+		setupEnv.Clients.ValidatorClient,
+		setupEnv.TokenID,
+		evm,
+		hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(),
+		setupEnv.TokenID.String(),
+		fmt.Sprint(mintAmount),
+		targetAsset,
+	)
 
 	// Step 4.1 - Get the consensus timestamp of the transfer
 	tx, err := setupEnv.Clients.MirrorNode.GetSuccessfulTransaction(hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String())
 	if err != nil {
 		t.Fatal("failed to get successful transaction", err)
 	}
+
 	nanos, err := timestamp.FromString(tx.ConsensusTimestamp)
 	if err != nil {
 		t.Fatal("failed to parse consensus timestamp", err)
 	}
+
 	ts := timestamp.FromNanos(nanos)
 
 	// Step 5 - Submit Mint transaction
-	txHash := submitMintTransaction(evm, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), transactionData, common.HexToAddress(targetAsset), t)
+	txHash := submit.MintTransaction(t, evm, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), transactionData, common.HexToAddress(targetAsset))
 
 	// Step 6 - Wait for transaction to be mined
-	waitForTransaction(evm, txHash, t)
+	submit.WaitForTransaction(t, evm, txHash)
 
 	// Step 7 - Validate Token balances
-	verifyWrappedAssetBalance(evm, targetAsset, big.NewInt(mintAmount), wrappedBalanceBefore, evm.Receiver, t)
+	verify.WrappedAssetBalance(t, evm, targetAsset, big.NewInt(mintAmount), wrappedBalanceBefore, evm.Receiver)
 
 	// Step 8 - Verify Database records
-	expectedTxRecord := util.PrepareExpectedTransfer(
+	expectedTxRecord := expected.FungibleTransferRecord(
 		constants.HederaNetworkId,
 		chainId,
 		constants.HederaNetworkId,
@@ -219,12 +232,15 @@ func Test_E2E_Token_Transfer(t *testing.T) {
 		evm.Receiver.String(),
 		status.Completed,
 		setupEnv.Clients.Hedera.GetOperatorAccountID().String(),
-		entity.NanoTime{Time: ts})
+		entity.NanoTime{Time: ts},
+	)
+
 	// and:
-	expectedFeeRecord := util.PrepareExpectedFeeRecord(
+	expectedFeeRecord := expected.FeeRecord(
 		scheduledTxID,
 		scheduleID, fee,
-		hederahelper.ToMirrorNodeTransactionID(transactionResponse.TransactionID.String()))
+		hederahelper.ToMirrorNodeTransactionID(transactionResponse.TransactionID.String()),
+	)
 
 	authMsgBytes, err := auth_message.EncodeFungibleBytesFrom(
 		expectedTxRecord.SourceChainID,
@@ -240,12 +256,12 @@ func Test_E2E_Token_Transfer(t *testing.T) {
 	}
 
 	// 8. Wait for validators to update DB state after Scheduled TX is mined
-	time.Sleep(20 * time.Second)
+	time.Sleep(30 * time.Second)
 
 	// Step 9 - Verify Database Records
-	verifyTransferRecordAndSignatures(setupEnv.DbValidator, expectedTxRecord, authMsgBytes, receivedSignatures, t)
+	verify.TransferRecordAndSignatures(t, setupEnv.DbValidator, expectedTxRecord, authMsgBytes, receivedSignatures)
 	// and:
-	verifyFeeRecord(setupEnv.DbValidator, expectedFeeRecord, t)
+	verify.FeeRecord(t, setupEnv.DbValidator, expectedFeeRecord)
 }
 
 // Test_EVM_Hedera_HBAR recreates a real life situation of a user who wants to return a Hedera native HBARs from the EVM Network infrastructure. The wrapped HBARs on the EVM network(corresponding to the native Hedera Hashgraph's one) gets burned, then the locked HBARs on the Hedera bridge account get unlocked, forwarding them to the recipient account.
@@ -255,19 +271,19 @@ func Test_EVM_Hedera_HBAR(t *testing.T) {
 
 	chainId := setupEnv.Scenario.FirstEvmChainId
 	evm := setupEnv.Clients.EVM[chainId]
-	now = time.Now()
-	accountBalanceBefore := util.GetHederaAccountBalance(setupEnv.Clients.Hedera, setupEnv.Clients.Hedera.GetOperatorAccountID(), t)
+	now := time.Now()
+	accountBalanceBefore := fetch.HederaAccountBalance(t, setupEnv.Clients.Hedera, setupEnv.Clients.Hedera.GetOperatorAccountID())
 
-	targetAsset, err := setup.NativeToWrappedAsset(setupEnv.AssetMappings, constants.HederaNetworkId, chainId, constants.Hbar)
+	targetAsset, err := evmSetup.NativeToWrappedAsset(setupEnv.AssetMappings, constants.HederaNetworkId, chainId, constants.Hbar)
 	if err != nil {
 		t.Fatalf("Expecting Token [%s] is not supported. - Error: [%s]", constants.Hbar, err)
 	}
 
 	// 1. Calculate Expected Receive And Fee Amounts
-	expectedReceiveAmount, fee := calculateReceiverAndFeeAmounts(setupEnv, constants.Hbar, amount)
+	expectedReceiveAmount, fee := expected.ReceiverAndFeeAmounts(setupEnv.Clients.FeeCalculator, setupEnv.Clients.Distributor, constants.Hbar, amount)
 
 	// 2. Submit burn transaction to the bridge contract
-	burnTxReceipt, expectedRouterBurn := sendBurnEthTransaction(setupEnv.AssetMappings, evm, constants.Hbar, constants.HederaNetworkId, chainId, setupEnv.Clients.Hedera.GetOperatorAccountID().ToBytes(), amount, t)
+	burnTxReceipt, expectedRouterBurn := submit.BurnEthTransaction(t, setupEnv.AssetMappings, evm, constants.Hbar, constants.HederaNetworkId, chainId, setupEnv.Clients.Hedera.GetOperatorAccountID().ToBytes(), amount)
 
 	// 2.1 Get the block timestamp of burn event
 	block, err := evm.EVMClient.BlockByNumber(context.Background(), burnTxReceipt.BlockNumber)
@@ -277,19 +293,20 @@ func Test_EVM_Hedera_HBAR(t *testing.T) {
 	blockTimestamp := time.Unix(int64(block.Time()), 0).UTC()
 
 	// 3. Validate that the burn transaction went through and emitted the correct events
-	expectedId := validateBurnEvent(burnTxReceipt, expectedRouterBurn, t)
+	expectedId := verify.BurnEvent(t, burnTxReceipt, expectedRouterBurn)
 
 	// 4. Validate that a scheduled transaction was submitted
-	transactionID, scheduleID := validateSubmittedScheduledTx(setupEnv, constants.Hbar, generateMirrorNodeExpectedTransfersForBurnEvent(setupEnv, constants.Hbar, expectedReceiveAmount, fee), t)
+	expectedTransfers := expected.MirrorNodeExpectedTransfersForBurnEvent(setupEnv.Members, setupEnv.Clients.Hedera, setupEnv.BridgeAccount, constants.Hbar, expectedReceiveAmount, fee)
+	transactionID, scheduleID := verify.SubmittedScheduledTx(t, setupEnv.Clients.Hedera, setupEnv.Clients.MirrorNode, setupEnv.Members, constants.Hbar, expectedTransfers, now)
 
 	// 5. Validate Event Transaction ID retrieved from Validator API
-	validateEventTransactionIDFromValidatorAPI(setupEnv, expectedId, transactionID, t)
+	verify.EventTransactionIDFromValidatorAPI(t, setupEnv.Clients.ValidatorClient, expectedId, transactionID)
 
 	// 6. Validate that the balance of the receiver account (hedera) was changed with the correct amount
-	validateReceiverAccountBalance(setupEnv, uint64(expectedReceiveAmount), accountBalanceBefore, constants.Hbar, t)
+	verify.ReceiverAccountBalance(t, setupEnv.Clients.Hedera, uint64(expectedReceiveAmount), accountBalanceBefore, constants.Hbar, setupEnv.TokenID)
 
 	// 7. Prepare Expected Database Records
-	expectedBurnEventRecord := util.PrepareExpectedTransfer(
+	expectedBurnEventRecord := expected.FungibleTransferRecord(
 		chainId,
 		constants.HederaNetworkId,
 		constants.HederaNetworkId,
@@ -301,17 +318,18 @@ func Test_EVM_Hedera_HBAR(t *testing.T) {
 		setupEnv.Clients.Hedera.GetOperatorAccountID().String(),
 		status.Completed,
 		evm.Signer.Address(),
-		entity.NanoTime{Time: blockTimestamp})
+		entity.NanoTime{Time: blockTimestamp},
+	)
 	// and:
-	expectedFeeRecord := util.PrepareExpectedFeeRecord(transactionID, scheduleID, fee, expectedId)
+	expectedFeeRecord := expected.FeeRecord(transactionID, scheduleID, fee, expectedId)
 
 	// 8. Wait for validators to update DB state after Scheduled TX is mined
 	time.Sleep(30 * time.Second)
 
 	// 9. Validate Database Records
-	verifyTransferRecord(setupEnv.DbValidator, expectedBurnEventRecord, t)
+	verify.TransferRecord(t, setupEnv.DbValidator, expectedBurnEventRecord)
 	// and:
-	verifyFeeRecord(setupEnv.DbValidator, expectedFeeRecord, t)
+	verify.FeeRecord(t, setupEnv.DbValidator, expectedFeeRecord)
 }
 
 // Test_EVM_Hedera_Token recreates a real life situation of a user who wants to return a Hedera native token from the EVM Network infrastructure. The wrapped token on the EVM network(corresponding to the native Hedera one) gets burned, then the amount gets unlocked on the Hedera bridge account, forwarding it to the recipient account.
@@ -321,27 +339,19 @@ func Test_EVM_Hedera_Token(t *testing.T) {
 
 	chainId := setupEnv.Scenario.FirstEvmChainId
 	evm := setupEnv.Clients.EVM[chainId]
-	now = time.Now()
-	accountBalanceBefore := util.GetHederaAccountBalance(setupEnv.Clients.Hedera, setupEnv.Clients.Hedera.GetOperatorAccountID(), t)
+	now := time.Now()
+	accountBalanceBefore := fetch.HederaAccountBalance(t, setupEnv.Clients.Hedera, setupEnv.Clients.Hedera.GetOperatorAccountID())
 
-	targetAsset, err := setup.NativeToWrappedAsset(setupEnv.AssetMappings, constants.HederaNetworkId, chainId, setupEnv.TokenID.String())
+	targetAsset, err := evmSetup.NativeToWrappedAsset(setupEnv.AssetMappings, constants.HederaNetworkId, chainId, setupEnv.TokenID.String())
 	if err != nil {
 		t.Fatalf("Expecting Token [%s] is not supported. - Error: [%s]", setupEnv.TokenID.String(), err)
 	}
 
 	// 1. Calculate Expected Receive Amount
-	expectedReceiveAmount, fee := calculateReceiverAndFeeAmounts(setupEnv, setupEnv.TokenID.String(), amount)
+	expectedReceiveAmount, fee := expected.ReceiverAndFeeAmounts(setupEnv.Clients.FeeCalculator, setupEnv.Clients.Distributor, setupEnv.TokenID.String(), amount)
 
 	// 2. Submit burn transaction to the bridge contract
-	burnTxReceipt, expectedRouterBurn := sendBurnEthTransaction(
-		setupEnv.AssetMappings,
-		evm,
-		setupEnv.TokenID.String(),
-		constants.HederaNetworkId,
-		chainId,
-		setupEnv.Clients.Hedera.GetOperatorAccountID().ToBytes(),
-		amount,
-		t)
+	burnTxReceipt, expectedRouterBurn := submit.BurnEthTransaction(t, setupEnv.AssetMappings, evm, setupEnv.TokenID.String(), constants.HederaNetworkId, chainId, setupEnv.Clients.Hedera.GetOperatorAccountID().ToBytes(), amount)
 
 	// 2.1 Get the block timestamp of burn event
 	block, err := evm.EVMClient.BlockByNumber(context.Background(), burnTxReceipt.BlockNumber)
@@ -351,19 +361,20 @@ func Test_EVM_Hedera_Token(t *testing.T) {
 	blockTimestamp := time.Unix(int64(block.Time()), 0).UTC()
 
 	// 3. Validate that the burn transaction went through and emitted the correct events
-	expectedId := validateBurnEvent(burnTxReceipt, expectedRouterBurn, t)
+	expectedId := verify.BurnEvent(t, burnTxReceipt, expectedRouterBurn)
 
 	// 4. Validate that a scheduled transaction was submitted
-	transactionID, scheduleID := validateSubmittedScheduledTx(setupEnv, setupEnv.TokenID.String(), generateMirrorNodeExpectedTransfersForBurnEvent(setupEnv, setupEnv.TokenID.String(), expectedReceiveAmount, fee), t)
+	expectedTransfers := expected.MirrorNodeExpectedTransfersForBurnEvent(setupEnv.Members, setupEnv.Clients.Hedera, setupEnv.BridgeAccount, setupEnv.TokenID.String(), expectedReceiveAmount, fee)
+	transactionID, scheduleID := verify.SubmittedScheduledTx(t, setupEnv.Clients.Hedera, setupEnv.Clients.MirrorNode, setupEnv.Members, setupEnv.TokenID.String(), expectedTransfers, now)
 
 	// 5. Validate Event Transaction ID retrieved from Validator API
-	validateEventTransactionIDFromValidatorAPI(setupEnv, expectedId, transactionID, t)
+	verify.EventTransactionIDFromValidatorAPI(t, setupEnv.Clients.ValidatorClient, expectedId, transactionID)
 
 	// 6. Validate that the balance of the receiver account (hedera) was changed with the correct amount
-	validateReceiverAccountBalance(setupEnv, uint64(expectedReceiveAmount), accountBalanceBefore, setupEnv.TokenID.String(), t)
+	verify.ReceiverAccountBalance(t, setupEnv.Clients.Hedera, uint64(expectedReceiveAmount), accountBalanceBefore, setupEnv.TokenID.String(), setupEnv.TokenID)
 
 	// 7. Prepare Expected Database Records
-	expectedBurnEventRecord := util.PrepareExpectedTransfer(
+	expectedBurnEventRecord := expected.FungibleTransferRecord(
 		chainId,
 		constants.HederaNetworkId,
 		constants.HederaNetworkId,
@@ -375,17 +386,18 @@ func Test_EVM_Hedera_Token(t *testing.T) {
 		setupEnv.Clients.Hedera.GetOperatorAccountID().String(),
 		status.Completed,
 		evm.Signer.Address(),
-		entity.NanoTime{Time: blockTimestamp})
+		entity.NanoTime{Time: blockTimestamp},
+	)
 	// and:
-	expectedFeeRecord := util.PrepareExpectedFeeRecord(transactionID, scheduleID, fee, expectedId)
+	expectedFeeRecord := expected.FeeRecord(transactionID, scheduleID, fee, expectedId)
 
 	// 8. Wait for validators to update DB state after Scheduled TX is mined
 	time.Sleep(30 * time.Second)
 
 	// 9. Validate Database Records
-	verifyTransferRecord(setupEnv.DbValidator, expectedBurnEventRecord, t)
+	verify.TransferRecord(t, setupEnv.DbValidator, expectedBurnEventRecord)
 	// and:
-	verifyFeeRecord(setupEnv.DbValidator, expectedFeeRecord, t)
+	verify.FeeRecord(t, setupEnv.DbValidator, expectedFeeRecord)
 }
 
 // Test_EVM_Hedera_Native_Token recreates a real life situation of a user who wants to bridge an EVM native token to the Hedera infrastructure. A new wrapped token (corresponding to the native EVM one) gets minted to the bridge account, then gets transferred to the recipient account.
@@ -396,16 +408,25 @@ func Test_EVM_Hedera_Native_Token(t *testing.T) {
 
 	chainId := setupEnv.Scenario.FirstEvmChainId
 	evm := setupEnv.Clients.EVM[chainId]
-	now = time.Now()
-	bridgeAccountBalanceBefore := util.GetHederaAccountBalance(setupEnv.Clients.Hedera, setupEnv.BridgeAccount, t)
-	receiverAccountBalanceBefore := util.GetHederaAccountBalance(setupEnv.Clients.Hedera, setupEnv.Clients.Hedera.GetOperatorAccountID(), t)
-	targetAsset, err := setup.NativeToWrappedAsset(setupEnv.AssetMappings, chainId, constants.HederaNetworkId, setupEnv.NativeEvmToken)
+	now := time.Now()
+
+	bridgeAccountBalanceBefore := fetch.HederaAccountBalance(t, setupEnv.Clients.Hedera, setupEnv.BridgeAccount)
+	receiverAccountBalanceBefore := fetch.HederaAccountBalance(t, setupEnv.Clients.Hedera, setupEnv.Clients.Hedera.GetOperatorAccountID())
+
+	targetAsset, err := evmSetup.NativeToWrappedAsset(setupEnv.AssetMappings, chainId, constants.HederaNetworkId, setupEnv.NativeEvmToken)
 	if err != nil {
 		t.Fatal(err)
 	}
 
+	// feeData, err := evm.RouterContract.TokenFeeData(nil, common.HexToAddress(setupEnv.NativeEvmToken))
+	// if err != nil {
+	// 	t.Fatalf("failed to get token fee data: %v", err)
+	// }
+
+	// fee := amount * (feeData.ServiceFeePercentage.Int64() / constants.FeeMaxPercentage)
+
 	// Step 2: Submit Lock Txn from a deployed smart contract
-	receipt, expectedLockEventLog := sendLockEthTransaction(evm, setupEnv.NativeEvmToken, constants.HederaNetworkId, setupEnv.Clients.Hedera.GetOperatorAccountID().ToBytes(), amount, t)
+	receipt, expectedLockEventLog := submit.LockEthTransaction(t, evm, setupEnv.NativeEvmToken, constants.HederaNetworkId, setupEnv.Clients.Hedera.GetOperatorAccountID().ToBytes(), amount)
 
 	// Step 2.1 - Get the block timestamp of lock event
 	block, err := evm.EVMClient.BlockByNumber(context.Background(), receipt.BlockNumber)
@@ -415,10 +436,10 @@ func Test_EVM_Hedera_Native_Token(t *testing.T) {
 	blockTimestamp := time.Unix(int64(block.Time()), 0).UTC()
 
 	// Step 3: Validate Lock Event was emitted with correct data
-	lockEventId := validateLockEvent(receipt, expectedLockEventLog, t)
+	lockEventId := verify.LockEvent(t, receipt, expectedLockEventLog)
 
 	bridgedAmount := new(big.Int).Sub(expectedLockEventLog.Amount, expectedLockEventLog.ServiceFee)
-	expectedAmount, err := removeDecimals(bridgedAmount.Int64(), common.HexToAddress(setupEnv.NativeEvmToken), evm)
+	expectedAmount, err := compat.RemoveDecimals(bridgedAmount.Int64(), common.HexToAddress(setupEnv.NativeEvmToken), evm)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -432,13 +453,13 @@ func Test_EVM_Hedera_Native_Token(t *testing.T) {
 	}
 
 	// Step 4: Validate that a scheduled token mint txn was submitted successfully
-	bridgeMintTransactionID, bridgeMintScheduleID := validateScheduledMintTx(setupEnv, setupEnv.BridgeAccount, setupEnv.TokenID.String(), mintTransfer, t)
+	bridgeMintTransactionID, bridgeMintScheduleID := verify.ScheduledMintTx(t, setupEnv.Clients.Hedera, setupEnv.Clients.MirrorNode, setupEnv.BridgeAccount, setupEnv.TokenID.String(), mintTransfer, now)
 
 	// Wait for validators to update DB state after Scheduled TX is mined
 	time.Sleep(30 * time.Second)
 
 	// Step 5: Validate that Database statuses were changed correctly
-	expectedLockEventRecord := util.PrepareExpectedTransfer(
+	expectedLockEventRecord := expected.FungibleTransferRecord(
 		chainId,
 		constants.HederaNetworkId,
 		chainId,
@@ -450,33 +471,38 @@ func Test_EVM_Hedera_Native_Token(t *testing.T) {
 		setupEnv.Clients.Hedera.GetOperatorAccountID().String(),
 		status.Completed,
 		evm.Signer.Address(),
-		entity.NanoTime{Time: blockTimestamp})
+		entity.NanoTime{Time: blockTimestamp},
+	)
 
-	expectedScheduleMintRecord := &entity.Schedule{
-		TransactionID: bridgeMintTransactionID,
-		ScheduleID:    bridgeMintScheduleID,
-		Operation:     schedule.MINT,
-		Status:        status.Completed,
-		TransferID: sql.NullString{
+	expectedScheduleMintRecord := expected.ScheduleRecord(
+		bridgeMintTransactionID,
+		bridgeMintScheduleID,
+		schedule.MINT,
+		false,
+		status.Completed,
+		sql.NullString{
 			String: lockEventId,
 			Valid:  true,
 		},
-	}
+	)
 
 	// Wait for validators to update DB state after Scheduled TX is mined
 	time.Sleep(30 * time.Second)
 
 	// Step 6: Verify that records have been created successfully
-	verifyTransferRecord(setupEnv.DbValidator, expectedLockEventRecord, t)
-	verifyScheduleRecord(setupEnv.DbValidator, expectedScheduleMintRecord, t)
+	verify.TransferRecord(t, setupEnv.DbValidator, expectedLockEventRecord)
+	verify.ScheduleRecord(t, setupEnv.DbValidator, expectedScheduleMintRecord)
 
 	// Step 7: Validate that a scheduled transfer txn was submitted successfully
-	bridgeTransferTransactionID, bridgeTransferScheduleID := validateScheduledTx(
-		setupEnv,
+	bridgeTransferTransactionID, bridgeTransferScheduleID := verify.ScheduledTx(
+		t,
+		setupEnv.Clients.Hedera,
+		setupEnv.Clients.MirrorNode,
 		setupEnv.Clients.Hedera.GetOperatorAccountID(),
 		setupEnv.TokenID.String(),
-		generateMirrorNodeExpectedTransfersForLockEvent(setupEnv, targetAsset, expectedAmount),
-		t)
+		expected.MirrorNodeExpectedTransfersForLockEvent(setupEnv.Clients.Hedera, setupEnv.BridgeAccount, targetAsset, expectedAmount),
+		now,
+	)
 
 	// Wait for validators to update DB state after Scheduled TX is mined
 	time.Sleep(30 * time.Second)
@@ -494,53 +520,16 @@ func Test_EVM_Hedera_Native_Token(t *testing.T) {
 		},
 	}
 
-	verifyScheduleRecord(setupEnv.DbValidator, expectedScheduleTransferRecord, t)
+	verify.ScheduleRecord(t, setupEnv.DbValidator, expectedScheduleTransferRecord)
 	// Step 9: Validate Treasury(BridgeAccount) Balance and Receiver Balance
-	validateAccountBalance(setupEnv, setupEnv.BridgeAccount, 0, bridgeAccountBalanceBefore, targetAsset, t)
-	validateAccountBalance(setupEnv, setupEnv.Clients.Hedera.GetOperatorAccountID(), uint64(expectedAmount), receiverAccountBalanceBefore, targetAsset, t)
-}
-
-func removeDecimals(amount int64, asset common.Address, evm setup.EVMUtils) (int64, error) {
-	evmAsset, err := wtoken.NewWtoken(asset, evm.EVMClient)
-	if err != nil {
-		return 0, err
-	}
-
-	decimals, err := evmAsset.Decimals(nil)
-	if err != nil {
-		return 0, err
-	}
-
-	adaptation := decimals - 8
-	if adaptation > 0 {
-		adapted := amount / int64(math.Pow10(int(adaptation)))
-		return adapted, nil
-	}
-	return amount, nil
-}
-
-func addDecimals(amount int64, asset common.Address, evm setup.EVMUtils) (int64, error) {
-	evmAsset, err := wtoken.NewWtoken(asset, evm.EVMClient)
-	if err != nil {
-		return 0, err
-	}
-
-	decimals, err := evmAsset.Decimals(nil)
-	if err != nil {
-		return 0, err
-	}
-	adaptation := decimals - 8
-	if adaptation > 0 {
-		adapted := amount * int64(math.Pow10(int(adaptation)))
-		return adapted, nil
-	}
-	return amount, nil
+	verify.AccountBalance(t, setupEnv.Clients.Hedera, setupEnv.BridgeAccount, 0, bridgeAccountBalanceBefore, targetAsset)
+	verify.AccountBalance(t, setupEnv.Clients.Hedera, setupEnv.Clients.Hedera.GetOperatorAccountID(), uint64(expectedAmount), receiverAccountBalanceBefore, targetAsset)
 }
 
 // Test_E2E_Hedera_EVM_Native_Token recreates a real life situation of a user who wants to bridge a Hedera wrapped token to the EVM Native Network infrastructure. The wrapped token on the EVM network(corresponding to the native Hedera Hashgraph's one) gets minted, then transferred to the recipient account on the EVM network.
 func Test_E2E_Hedera_EVM_Native_Token(t *testing.T) {
 	setupEnv := setup.Load()
-	now = time.Now()
+	now := time.Now()
 
 	chainId := setupEnv.Scenario.FirstEvmChainId
 	evm := setupEnv.Clients.EVM[chainId]
@@ -548,7 +537,7 @@ func Test_E2E_Hedera_EVM_Native_Token(t *testing.T) {
 	unlockAmount := int64(10) // Amount, which converted to 18 decimals is 100000000000 (100 gwei)
 
 	// Step 1 - Verify the transfer of HTS to the Bridge Account
-	wrappedAsset, err := setup.NativeToWrappedAsset(setupEnv.AssetMappings, chainId, constants.HederaNetworkId, setupEnv.NativeEvmToken)
+	wrappedAsset, err := evmSetup.NativeToWrappedAsset(setupEnv.AssetMappings, chainId, constants.HederaNetworkId, setupEnv.NativeEvmToken)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -558,12 +547,13 @@ func Test_E2E_Hedera_EVM_Native_Token(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	expectedSubmitUnlockAmount, err := addDecimals(unlockAmount, common.HexToAddress(setupEnv.NativeEvmToken), evm)
+	expectedSubmitUnlockAmount, err := compat.AddDecimals(unlockAmount, common.HexToAddress(setupEnv.NativeEvmToken), evm)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	transactionResponse, nativeBalanceBefore := verifyTokenTransferToBridgeAccount(setupEnv, setupEnv.NativeEvmToken, tokenID, evm, memo, evm.Receiver, unlockAmount, t)
+	transactionResponse, nativeBalanceBefore := verify.TokenTransferToBridgeAccount(t, setupEnv.Clients.Hedera, setupEnv.BridgeAccount, setupEnv.NativeEvmToken, tokenID, evm, memo, evm.Receiver, unlockAmount)
+
 	burnTransfer := []transaction.Transfer{
 		{
 			Account: setupEnv.BridgeAccount.String(),
@@ -573,22 +563,23 @@ func Test_E2E_Hedera_EVM_Native_Token(t *testing.T) {
 	}
 
 	// Step 2 - Verify the submitted topic messages
-	receivedSignatures := verifyTopicMessages(setupEnv, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), t)
+	receivedSignatures := verify.TopicMessages(t, setupEnv.Clients.Hedera, setupEnv.TopicID, setupEnv.Scenario.ExpectedValidatorsCount, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String())
 
 	// Step 3 - Validate burn scheduled transaction
-	burnTransactionID, burnScheduleID := validateScheduledBurnTx(setupEnv, setupEnv.BridgeAccount, setupEnv.TokenID.String(), burnTransfer, t)
+	burnTransactionID, burnScheduleID := verify.ScheduledBurnTx(t, setupEnv.Clients.Hedera, setupEnv.Clients.MirrorNode, setupEnv.BridgeAccount, setupEnv.TokenID.String(), burnTransfer, now)
 
 	time.Sleep(30 * time.Second)
 
 	// Step 4 - Verify Transfer retrieved from Validator API
-	transactionData := verifyFungibleTransferFromValidatorAPI(
-		setupEnv,
+	transactionData := verify.FungibleTransferFromValidatorAPI(
+		t,
+		setupEnv.Clients.ValidatorClient,
+		setupEnv.TokenID,
 		evm,
 		hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(),
 		setupEnv.NativeEvmToken,
 		fmt.Sprint(expectedSubmitUnlockAmount),
 		setupEnv.NativeEvmToken,
-		t,
 	)
 
 	// Step 4.1 - Get the consensus timestamp of the transfer
@@ -603,21 +594,19 @@ func Test_E2E_Hedera_EVM_Native_Token(t *testing.T) {
 	ts := timestamp.FromNanos(nanos)
 
 	// Step 5 - Submit Unlock transaction
-	txHash := submitUnlockWithFeeTransaction(evm, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), transactionData, common.HexToAddress(setupEnv.NativeEvmToken), t)
+	txHash := submit.UnlockWithFeeTransaction(t, evm, hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(), transactionData, common.HexToAddress(setupEnv.NativeEvmToken))
 
 	// Step 6 - Wait for transaction to be mined
-	waitForTransaction(evm, txHash, t)
+	submit.WaitForTransaction(t, evm, txHash)
 
 	feeAmount, _ := strconv.ParseInt(transactionData.Fee, 10, 64)
 	expectedUnlockedAmount := expectedSubmitUnlockAmount - feeAmount
 
-	// expectedUnlockedAmount, feeAmount := calculateExpectedUnlockAmount(setupEnv, evm, setupEnv.NativeEvmToken, expectedSubmitUnlockAmount, t)
-
 	// Step 7 - Validate Token balances
-	verifyWrappedAssetBalance(evm, setupEnv.NativeEvmToken, big.NewInt(expectedUnlockedAmount), nativeBalanceBefore, evm.Receiver, t)
+	verify.WrappedAssetBalance(t, evm, setupEnv.NativeEvmToken, big.NewInt(expectedUnlockedAmount), nativeBalanceBefore, evm.Receiver)
 
 	// Step 8 - Verify Database records
-	expectedTxRecord := util.PrepareExpectedTransfer(
+	expectedTxRecord := expected.FungibleTransferRecord(
 		constants.HederaNetworkId,
 		chainId,
 		chainId,
@@ -629,19 +618,21 @@ func Test_E2E_Hedera_EVM_Native_Token(t *testing.T) {
 		evm.Receiver.String(),
 		status.Completed,
 		setupEnv.Clients.Hedera.GetOperatorAccountID().String(),
-		entity.NanoTime{Time: ts})
+		entity.NanoTime{Time: ts},
+	)
 
 	// Step 8: Validate that database statuses were updated correctly for the Schedule Burn
-	expectedScheduleBurnRecord := &entity.Schedule{
-		TransactionID: burnTransactionID,
-		ScheduleID:    burnScheduleID,
-		Operation:     schedule.BURN,
-		Status:        status.Completed,
-		TransferID: sql.NullString{
+	expectedScheduleBurnRecord := expected.ScheduleRecord(
+		burnTransactionID,
+		burnScheduleID,
+		schedule.BURN,
+		false,
+		status.Completed,
+		sql.NullString{
 			String: hederahelper.FromHederaTransactionID(transactionResponse.TransactionID).String(),
 			Valid:  true,
 		},
-	}
+	)
 
 	authMsgBytes, err := auth_message.EncodeFungibleBytesFromWithFee(
 		expectedTxRecord.SourceChainID,
@@ -658,9 +649,9 @@ func Test_E2E_Hedera_EVM_Native_Token(t *testing.T) {
 	}
 
 	// Step 9 - Verify Database Records
-	verifyTransferRecordAndSignatures(setupEnv.DbValidator, expectedTxRecord, authMsgBytes, receivedSignatures, t)
+	verify.TransferRecordAndSignatures(t, setupEnv.DbValidator, expectedTxRecord, authMsgBytes, receivedSignatures)
 	// and
-	verifyScheduleRecord(setupEnv.DbValidator, expectedScheduleBurnRecord, t)
+	verify.ScheduleRecord(t, setupEnv.DbValidator, expectedScheduleBurnRecord)
 }
 
 // Test_EVM_Native_to_EVM_Token recreates a real life situation of a user who wants to bridge an EVM native token to another EVM chain.
@@ -673,13 +664,19 @@ func Test_EVM_Native_to_EVM_Token(t *testing.T) {
 	evm := setupEnv.Clients.EVM[chainId]
 	now = time.Now()
 	targetChainID := setupEnv.Scenario.SecondEvmChainId
-	wrappedAsset, err := setup.NativeToWrappedAsset(setupEnv.AssetMappings, chainId, targetChainID, setupEnv.NativeEvmToken)
+
+	wrappedAsset, err := evmSetup.NativeToWrappedAsset(setupEnv.AssetMappings, chainId, targetChainID, setupEnv.NativeEvmToken)
 	if err != nil {
 		t.Fatal(err)
 	}
+	// feeData, err := evm.RouterContract.TokenFeeData(nil, common.HexToAddress(setupEnv.NativeEvmToken))
+	// if err != nil {
+	// 	t.Fatalf("failed to get token fee data: %v", err)
+	// }
+	// fee := amount * (feeData.ServiceFeePercentage.Int64() / constants.FeeMaxPercentage)
 
 	wrappedEvm := setupEnv.Clients.EVM[targetChainID]
-	wrappedInstance, err := setup.InitAssetContract(wrappedAsset, wrappedEvm.EVMClient)
+	wrappedInstance, err := evmSetup.InitAssetContract(wrappedAsset, wrappedEvm.EVMClient)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -690,7 +687,7 @@ func Test_EVM_Native_to_EVM_Token(t *testing.T) {
 	}
 
 	// Step 2 - Submit Lock Txn from a deployed smart contract
-	receipt, expectedLockEventLog := sendLockEthTransaction(evm, setupEnv.NativeEvmToken, targetChainID, evm.Receiver.Bytes(), amount, t)
+	receipt, expectedLockEventLog := submit.LockEthTransaction(t, evm, setupEnv.NativeEvmToken, targetChainID, evm.Receiver.Bytes(), amount)
 
 	expectedAmount := new(big.Int).Sub(expectedLockEventLog.Amount, expectedLockEventLog.ServiceFee)
 
@@ -702,27 +699,27 @@ func Test_EVM_Native_to_EVM_Token(t *testing.T) {
 	blockTimestamp := time.Unix(int64(block.Time()), 0).UTC()
 
 	// Step 3 - Validate Lock Event was emitted with correct data
-	lockEventId := validateLockEvent(receipt, expectedLockEventLog, t)
+	lockEventId := verify.LockEvent(t, receipt, expectedLockEventLog)
 
 	// Step 4 - Verify the submitted topic messages
-	receivedSignatures := verifyTopicMessages(setupEnv, lockEventId, t)
+	receivedSignatures := verify.TopicMessages(t, setupEnv.Clients.Hedera, setupEnv.TopicID, setupEnv.Scenario.ExpectedValidatorsCount, lockEventId)
 
 	time.Sleep(30 * time.Second)
 
 	// Step 5 - Verify Transfer retrieved from Validator API
-	transactionData := verifyFungibleTransferFromValidatorAPI(setupEnv, evm, lockEventId, setupEnv.NativeEvmToken, expectedAmount.String(), wrappedAsset, t)
+	transactionData := verify.FungibleTransferFromValidatorAPI(t, setupEnv.Clients.ValidatorClient, setupEnv.TokenID, evm, lockEventId, setupEnv.NativeEvmToken, expectedAmount.String(), wrappedAsset)
 
 	// Step 6 - Submit Mint transaction
-	txHash := submitMintTransaction(wrappedEvm, lockEventId, transactionData, common.HexToAddress(wrappedAsset), t)
+	txHash := submit.MintTransaction(t, wrappedEvm, lockEventId, transactionData, common.HexToAddress(wrappedAsset))
 
 	// Step 7 - Wait for transaction to be mined
-	waitForTransaction(wrappedEvm, txHash, t)
+	submit.WaitForTransaction(t, wrappedEvm, txHash)
 
 	// Step 8 - Validate Token balances
-	verifyWrappedAssetBalance(wrappedEvm, wrappedAsset, expectedAmount, wrappedBalanceBefore, evm.Receiver, t)
+	verify.WrappedAssetBalance(t, wrappedEvm, wrappedAsset, expectedAmount, wrappedBalanceBefore, evm.Receiver)
 
 	// Step 9 - Prepare expected Transfer record
-	expectedLockEventRecord := util.PrepareExpectedTransfer(
+	expectedLockEventRecord := expected.FungibleTransferRecord(
 		chainId,
 		targetChainID,
 		chainId,
@@ -734,7 +731,8 @@ func Test_EVM_Native_to_EVM_Token(t *testing.T) {
 		evm.Receiver.String(),
 		status.Completed,
 		evm.Signer.Address(),
-		entity.NanoTime{Time: blockTimestamp})
+		entity.NanoTime{Time: blockTimestamp},
+	)
 
 	authMsgBytes, err := auth_message.EncodeFungibleBytesFrom(
 		expectedLockEventRecord.SourceChainID,
@@ -744,12 +742,13 @@ func Test_EVM_Native_to_EVM_Token(t *testing.T) {
 		expectedLockEventRecord.Receiver,
 		expectedAmount.String(),
 	)
+
 	if err != nil {
 		t.Fatalf("[%s] - Failed to encode the authorisation signature. Error: [%s]", expectedLockEventRecord.TransactionID, err)
 	}
 
 	// Step 10 - Verify Database Records
-	verifyTransferRecordAndSignatures(setupEnv.DbValidator, expectedLockEventRecord, authMsgBytes, receivedSignatures, t)
+	verify.TransferRecordAndSignatures(t, setupEnv.DbValidator, expectedLockEventRecord, authMsgBytes, receivedSignatures)
 }
 
 // Test_EVM_Wrapped_to_EVM_Token recreates a real life situation of a user who wants to bridge an EVM native token to another EVM chain.
@@ -761,14 +760,15 @@ func Test_EVM_Wrapped_to_EVM_Token(t *testing.T) {
 	chainId := setupEnv.Scenario.FirstEvmChainId
 	sourceChain := setupEnv.Scenario.SecondEvmChainId
 	wrappedEvm := setupEnv.Clients.EVM[sourceChain]
-	now = time.Now()
-	sourceAsset, err := setup.NativeToWrappedAsset(setupEnv.AssetMappings, chainId, sourceChain, setupEnv.NativeEvmToken)
+
+	sourceAsset, err := evmSetup.NativeToWrappedAsset(setupEnv.AssetMappings, chainId, sourceChain, setupEnv.NativeEvmToken)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	nativeEvm := setupEnv.Clients.EVM[chainId]
-	nativeInstance, err := setup.InitAssetContract(setupEnv.NativeEvmToken, nativeEvm.EVMClient)
+
+	nativeInstance, err := evmSetup.InitAssetContract(setupEnv.NativeEvmToken, nativeEvm.EVMClient)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -779,7 +779,7 @@ func Test_EVM_Wrapped_to_EVM_Token(t *testing.T) {
 	}
 
 	// Step 2 - Submit Lock Txn from a deployed smart contract
-	receipt, expectedLockEventLog := sendBurnEthTransaction(setupEnv.AssetMappings, wrappedEvm, setupEnv.NativeEvmToken, chainId, sourceChain, nativeEvm.Receiver.Bytes(), amount, t)
+	receipt, expectedLockEventLog := submit.BurnEthTransaction(t, setupEnv.AssetMappings, wrappedEvm, setupEnv.NativeEvmToken, chainId, sourceChain, nativeEvm.Receiver.Bytes(), amount)
 
 	// Step 2.1 - Get the block timestamp of the burn event
 	block, err := wrappedEvm.EVMClient.BlockByNumber(context.Background(), receipt.BlockNumber)
@@ -789,13 +789,13 @@ func Test_EVM_Wrapped_to_EVM_Token(t *testing.T) {
 	blockTimestamp := time.Unix(int64(block.Time()), 0).UTC()
 
 	// Step 3 - Validate Burn Event was emitted with correct data
-	burnEventId := validateBurnEvent(receipt, expectedLockEventLog, t)
+	burnEventId := verify.BurnEvent(t, receipt, expectedLockEventLog)
 
 	// Step 4 - Verify the submitted topic messages
-	receivedSignatures := verifyTopicMessages(setupEnv, burnEventId, t)
+	receivedSignatures := verify.TopicMessages(t, setupEnv.Clients.Hedera, setupEnv.TopicID, setupEnv.Scenario.ExpectedValidatorsCount, burnEventId)
 
 	// Step 5 - Verify Transfer retrieved from Validator API
-	transactionData := verifyFungibleTransferFromValidatorAPI(setupEnv, nativeEvm, burnEventId, setupEnv.NativeEvmToken, fmt.Sprint(amount), setupEnv.NativeEvmToken, t)
+	transactionData := verify.FungibleTransferFromValidatorAPI(t, setupEnv.Clients.ValidatorClient, setupEnv.TokenID, nativeEvm, burnEventId, setupEnv.NativeEvmToken, fmt.Sprint(amount), setupEnv.NativeEvmToken)
 
 	// Get fee amount from wrapped network Router
 	feeAmount, err := nativeEvm.RouterContract.FeeAmountFor(nil, new(big.Int).SetUint64(sourceChain), nativeEvm.Receiver, common.HexToAddress(setupEnv.NativeEvmToken), new(big.Int).SetInt64(amount))
@@ -807,18 +807,18 @@ func Test_EVM_Wrapped_to_EVM_Token(t *testing.T) {
 	transactionData.Fee = feeAmount.String()
 
 	// Step 6 - Submit Mint transaction
-	txHash := submitUnlockTransaction(nativeEvm, burnEventId, transactionData, common.HexToAddress(setupEnv.NativeEvmToken), t)
+	txHash := submit.UnlockTransaction(t, nativeEvm, burnEventId, transactionData, common.HexToAddress(setupEnv.NativeEvmToken))
 
 	// Step 7 - Wait for transaction to be mined
-	waitForTransaction(nativeEvm, txHash, t)
+	submit.WaitForTransaction(t, nativeEvm, txHash)
 
 	expectedUnlockedAmount := amount - feeAmount.Int64()
 
 	// Step 8 - Validate Token balances
-	verifyWrappedAssetBalance(nativeEvm, setupEnv.NativeEvmToken, big.NewInt(expectedUnlockedAmount), nativeBalanceBefore, nativeEvm.Receiver, t)
+	verify.WrappedAssetBalance(t, nativeEvm, setupEnv.NativeEvmToken, big.NewInt(expectedUnlockedAmount), nativeBalanceBefore, nativeEvm.Receiver)
 
 	// Step 9 - Prepare expected Transfer record
-	expectedLockEventRecord := util.PrepareExpectedTransfer(
+	expectedLockEventRecord := expected.FungibleTransferRecord(
 		sourceChain,
 		chainId,
 		chainId,
@@ -830,7 +830,8 @@ func Test_EVM_Wrapped_to_EVM_Token(t *testing.T) {
 		nativeEvm.Receiver.String(),
 		status.Completed,
 		wrappedEvm.Signer.Address(),
-		entity.NanoTime{Time: blockTimestamp})
+		entity.NanoTime{Time: blockTimestamp},
+	)
 
 	authMsgBytes, err := auth_message.EncodeFungibleBytesFrom(
 		expectedLockEventRecord.SourceChainID,
@@ -849,7 +850,7 @@ func Test_EVM_Wrapped_to_EVM_Token(t *testing.T) {
 	time.Sleep(30 * time.Second)
 
 	// Step 9 - Verify Database Records
-	verifyTransferRecordAndSignatures(setupEnv.DbValidator, expectedLockEventRecord, authMsgBytes, receivedSignatures, t)
+	verify.TransferRecordAndSignatures(t, setupEnv.DbValidator, expectedLockEventRecord, authMsgBytes, receivedSignatures)
 }
 
 // Test_Hedera_Native_EVM_NFT_Transfer recreates User who wants to portal a Hedera Native NFT to an EVM chain.
@@ -898,33 +899,42 @@ func Test_Hedera_Native_EVM_NFT_Transfer(t *testing.T) {
 	}
 
 	// Step 2 - Send the NFT Allowance for the Payer Account
-	_, err = sendNFTAllowance(setupEnv, nftID, setupEnv.Clients.Hedera.GetOperatorAccountID(), setupEnv.PayerAccount)
+	_, err = verify.SendNFTAllowance(setupEnv.Clients.Hedera, nftID, setupEnv.Clients.Hedera.GetOperatorAccountID(), setupEnv.PayerAccount)
 	if err != nil {
 		t.Fatalf("Failed to send Allowance for NFT [%s]. Error: [%s]", nftIDString, err)
 	}
 	signaturesStartTime := time.Now().UnixNano()
+
 	// Step 3 - Send the NFT transfer, including the fee to the Bridge Account
-	feeResponse, err := sendFeeForNFTToBridgeAccount(setupEnv, evm.Receiver, chainId, nftID, transferFee)
+	memo := fmt.Sprintf("%d-%s-%s", chainId, receiver, nftID.String())
+	feeResponse, err := submit.FeeForNFTToBridgeAccount(setupEnv.Clients.Hedera, setupEnv.BridgeAccount, memo, chainId, nftID, transferFee)
 	if err != nil {
 		t.Fatalf("Failed to send Fee and allowance for NFT transfer. Error: [%s]", err)
 	}
 	transactionID := hederahelper.FromHederaTransactionID(feeResponse.TransactionID).String()
 
 	// Step 4 - Validate that a scheduled NFT transaction to the bridge account was submitted by a validator
-	validateScheduledNftTransfer(setupEnv, nftToken, serialNumber, t)
+	verify.ScheduledNftTransfer(t, setupEnv.Clients.Hedera, setupEnv.Clients.MirrorNode, setupEnv.BridgeAccount, nftToken, serialNumber)
 
 	// Step 5 - Verify the submitted topic messages
-	receivedSignatures := verifyTopicMessagesWithStartTime(setupEnv, hederahelper.FromHederaTransactionID(feeResponse.TransactionID).String(), signaturesStartTime, t)
+	receivedSignatures := verify.TopicMessagesWithStartTime(t, setupEnv.Clients.Hedera, setupEnv.TopicID, setupEnv.Scenario.ExpectedValidatorsCount, hederahelper.FromHederaTransactionID(feeResponse.TransactionID).String(), signaturesStartTime)
 
 	// Step 6 - Validate members fee scheduled transaction
-	scheduledTxID, scheduleID := validateMembersScheduledTxs(setupEnv, constants.Hbar, generateMirrorNodeExpectedTransfersForHederaTransfer(setupEnv, constants.Hbar, validatorsFee), t)
+	expectedTransfers := expected.MirrorNodeExpectedTransfersForHederaTransfer(setupEnv.Members, setupEnv.BridgeAccount, constants.Hbar, validatorsFee)
+	scheduledTxID, scheduleID := verify.MembersScheduledTxs(t, setupEnv.Clients.Hedera, setupEnv.Clients.MirrorNode, setupEnv.Members, constants.Hbar, expectedTransfers, now)
 
 	// Step 7 - Verify Non-Fungible Transfer retrieved from Validator API
-	transactionData := verifyNonFungibleTransferFromValidatorAPI(
-		setupEnv,
+	transactionData := verify.NonFungibleTransferFromValidatorAPI(
+		t,
+		setupEnv.Clients.ValidatorClient,
+		setupEnv.TokenID,
 		evm,
 		transactionID,
-		nftToken, string(decodedMetadata), serialNumber, targetAsset, t)
+		nftToken,
+		string(decodedMetadata),
+		serialNumber,
+		targetAsset,
+	)
 
 	// Step 7.1 - Get the consensus timestamp of the transfer
 	tx, err := setupEnv.Clients.MirrorNode.GetSuccessfulTransaction(transactionID)
@@ -938,13 +948,13 @@ func Test_Hedera_Native_EVM_NFT_Transfer(t *testing.T) {
 	ts := timestamp.FromNanos(nanos)
 
 	// Step 8 - Submit Mint ERC-721 transaction
-	txHash := submitMintERC721Transaction(evm, transactionID, transactionData, t)
+	txHash := submit.MintERC721Transaction(t, evm, transactionID, transactionData)
 
 	// Step 9 - Wait for transaction to be mined
-	waitForTransaction(evm, txHash, t)
+	submit.WaitForTransaction(t, evm, txHash)
 
 	// Step 10 - Validate EVM TokenId
-	verifyERC721TokenId(evm.EVMClient, targetAsset, serialNumber, receiver.String(), string(decodedMetadata), t)
+	verify.ERC721TokenId(t, evm.EVMClient, targetAsset, serialNumber, receiver.String(), string(decodedMetadata))
 
 	// Step 11 - Verify Database records
 	expectedTxRecord := &entity.Transfer{
@@ -966,7 +976,7 @@ func Test_Hedera_Native_EVM_NFT_Transfer(t *testing.T) {
 		Timestamp:     entity.NanoTime{Time: ts},
 	}
 	// and:
-	expectedFeeRecord := util.PrepareExpectedFeeRecord(
+	expectedFeeRecord := expected.FeeRecord(
 		scheduledTxID,
 		scheduleID, validatorsFee,
 		transactionID)
@@ -1000,11 +1010,11 @@ func Test_Hedera_Native_EVM_NFT_Transfer(t *testing.T) {
 	time.Sleep(20 * time.Second)
 
 	// Step 12 - Verify Database Records
-	verifyTransferRecordAndSignatures(setupEnv.DbValidator, expectedTxRecord, authMsgBytes, receivedSignatures, t)
+	verify.TransferRecordAndSignatures(t, setupEnv.DbValidator, expectedTxRecord, authMsgBytes, receivedSignatures)
 	// and:
-	verifyFeeRecord(setupEnv.DbValidator, expectedFeeRecord, t)
+	verify.FeeRecord(t, setupEnv.DbValidator, expectedFeeRecord)
 	// and:
-	verifyScheduleRecord(setupEnv.DbValidator, expectedScheduleTransferRecord, t)
+	verify.ScheduleRecord(t, setupEnv.DbValidator, expectedScheduleTransferRecord)
 }
 
 // Test_Hedera_EVM_BurnERC721_Transfer recreates User who wants to portal back a Hedera Native NFT from an EVM chain.
@@ -1023,10 +1033,10 @@ func Test_Hedera_EVM_BurnERC721_Transfer(t *testing.T) {
 	}
 
 	// 1. Validate that NFT spender is the bridge account
-	validateNftOwner(setupEnv, nftToken, serialNumber, setupEnv.BridgeAccount, t)
+	verify.NftOwner(t, setupEnv.Clients.Hedera, nftToken, serialNumber, setupEnv.BridgeAccount)
 
 	// 2. Submit burnERC721 transaction to the bridge contract
-	burnTxReceipt, expectedRouterBurnERC721 := sendBurnERC721Transaction(evm, wrappedAsset, constants.HederaNetworkId, setupEnv.Clients.Hedera.GetOperatorAccountID().ToBytes(), serialNumber, t)
+	burnTxReceipt, expectedRouterBurnERC721 := submit.BurnERC721Transaction(t, evm, wrappedAsset, constants.HederaNetworkId, setupEnv.Clients.Hedera.GetOperatorAccountID().ToBytes(), serialNumber)
 
 	// 2.1 - Get the block timestamp of the burn event
 	block, err := evm.EVMClient.BlockByNumber(context.Background(), burnTxReceipt.BlockNumber)
@@ -1036,16 +1046,16 @@ func Test_Hedera_EVM_BurnERC721_Transfer(t *testing.T) {
 	blockTimestamp := time.Unix(int64(block.Time()), 0).UTC()
 
 	// 3. Validate that the burn ERC-721 transaction went through and emitted the correct event
-	expectedTxId := validateBurnERC721Event(burnTxReceipt, expectedRouterBurnERC721, t)
+	expectedTxId := verify.BurnERC721Event(t, burnTxReceipt, expectedRouterBurnERC721)
 
 	// 4. Validate that a scheduled NFT transaction was submitted
-	scheduledTxID, scheduleID := validateScheduledNftAllowanceApprove(t, setupEnv, expectedTxId, blockTimestamp.UnixNano())
+	scheduledTxID, scheduleID := verify.ScheduledNftAllowanceApprove(t, setupEnv.Clients.Hedera, setupEnv.Clients.MirrorNode, setupEnv.BridgeAccount, setupEnv.PayerAccount, expectedTxId, blockTimestamp.UnixNano())
 
 	// 5. Validate Event Transaction ID retrieved from Validator API
-	validateEventTransactionIDFromValidatorAPI(setupEnv, expectedTxId, scheduledTxID, t)
+	verify.EventTransactionIDFromValidatorAPI(t, setupEnv.Clients.ValidatorClient, expectedTxId, scheduledTxID)
 
 	// 6. Validate that the NFT is allowed to the receiver account
-	validateNftSpender(t, setupEnv, nftToken, serialNumber, setupEnv.Clients.Hedera.GetOperatorAccountID())
+	verify.NftSpender(t, setupEnv.Clients.Hedera, nftToken, serialNumber, setupEnv.Clients.Hedera.GetOperatorAccountID())
 
 	// 7. Transfer the NFT to the receiver account
 	tx, err := hedera.NewTransferTransaction().
@@ -1063,7 +1073,7 @@ func Test_Hedera_EVM_BurnERC721_Transfer(t *testing.T) {
 	}
 
 	// 8. Validate that the NFT is transferred to the receiver account
-	validateNftOwner(setupEnv, nftToken, serialNumber, setupEnv.Clients.Hedera.GetOperatorAccountID(), t)
+	verify.NftOwner(t, setupEnv.Clients.Hedera, nftToken, serialNumber, setupEnv.Clients.Hedera.GetOperatorAccountID())
 
 	// 9. Prepare Expected Database Records
 	expectedTxRecord := &entity.Transfer{
@@ -1094,1316 +1104,11 @@ func Test_Hedera_EVM_BurnERC721_Transfer(t *testing.T) {
 	}
 
 	// 10. Wait for validators to update DB state after Scheduled TX is mined
-	time.Sleep(20 * time.Second)
+	time.Sleep(30 * time.Second)
 
 	// 11. Validate Database Records
-	verifyTransferRecord(setupEnv.DbValidator, expectedTxRecord, t)
+	verify.TransferRecord(t, setupEnv.DbValidator, expectedTxRecord)
 	// and:
-	verifyScheduleRecord(setupEnv.DbValidator, expectedScheduleTransferRecord, t)
-}
+	verify.ScheduleRecord(t, setupEnv.DbValidator, expectedScheduleTransferRecord)
 
-func validateNftOwner(setup *setup.Setup, tokenID string, serialNumber int64, expectedOwner hedera.AccountID, t *testing.T) {
-	nftID, err := hedera.NftIDFromString(fmt.Sprintf("%d@%s", serialNumber, tokenID))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	nftInfo, err := hedera.NewTokenNftInfoQuery().
-		SetNftID(nftID).
-		Execute(setup.Clients.Hedera)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(nftInfo) != 1 {
-		t.Fatalf("Invalid NFT Info [%s] length result. Result: [%v]", nftID.String(), nftInfo)
-	}
-
-	owner := nftInfo[0].AccountID
-	if owner != expectedOwner {
-		t.Fatalf("Invalid NftID [%s] owner. Expected [%s], actual [%s].", nftID.String(), expectedOwner, owner)
-	}
-}
-
-func validateNftSpender(t *testing.T, setup *setup.Setup, tokenID string, serialNumber int64, expectedSpender hedera.AccountID) {
-	tokenIdFromString, err := hedera.TokenIDFromString(tokenID)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	nftId := hedera.NftID{
-		TokenID:      tokenIdFromString,
-		SerialNumber: serialNumber,
-	}
-
-	nftInfo, err := hedera.NewTokenNftInfoQuery().
-		SetNftID(nftId).
-		Execute(setup.Clients.Hedera)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(nftInfo) != 1 {
-		t.Fatalf("Invalid NFT Info [%s] length result. Result: [%v]", nftId.String(), nftInfo)
-	}
-
-	spender := nftInfo[0].SpenderID
-	if spender != expectedSpender {
-		t.Fatalf("Invalid NftID [%s] spender. Expected [%s], actual [%s].", nftId.String(), expectedSpender, spender)
-	}
-}
-
-func validateReceiverAccountBalance(setup *setup.Setup, expectedReceiveAmount uint64, beforeHbarBalance hedera.AccountBalance, asset string, t *testing.T) {
-	afterHbarBalance := util.GetHederaAccountBalance(setup.Clients.Hedera, setup.Clients.Hedera.GetOperatorAccountID(), t)
-
-	var beforeTransfer uint64
-	var afterTransfer uint64
-
-	if asset == constants.Hbar {
-		beforeTransfer = uint64(beforeHbarBalance.Hbars.AsTinybar())
-		afterTransfer = uint64(afterHbarBalance.Hbars.AsTinybar())
-	} else {
-		beforeTransfer = beforeHbarBalance.Tokens.Get(setup.TokenID)
-		afterTransfer = afterHbarBalance.Tokens.Get(setup.TokenID)
-	}
-
-	if afterTransfer-beforeTransfer != expectedReceiveAmount {
-		t.Fatalf("[%s] Expected %s balance after - [%d], but was [%d]. Expected to receive [%d], but was [%d]", setup.Clients.Hedera.GetOperatorAccountID(), asset, beforeTransfer+expectedReceiveAmount, afterTransfer, expectedReceiveAmount, afterTransfer-beforeTransfer)
-	}
-}
-
-func validateAccountBalance(setup *setup.Setup, hederaID hedera.AccountID, expectedReceiveAmount uint64, beforeHbarBalance hedera.AccountBalance, asset string, t *testing.T) {
-	afterHbarBalance := util.GetHederaAccountBalance(setup.Clients.Hedera, hederaID, t)
-
-	tokenAsset, err := hedera.TokenIDFromString(asset)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	beforeTransfer := beforeHbarBalance.Tokens.Get(tokenAsset)
-	afterTransfer := afterHbarBalance.Tokens.Get(tokenAsset)
-
-	if afterTransfer-beforeTransfer != expectedReceiveAmount {
-		t.Fatalf("[%s] Expected %s balance after - [%d], but was [%d]. Expected to receive [%d], but was [%d]", setup.Clients.Hedera.GetOperatorAccountID(), asset, beforeTransfer+expectedReceiveAmount, afterTransfer, expectedReceiveAmount, afterTransfer-beforeTransfer)
-	}
-}
-
-func validateBurnEvent(txReceipt *types.Receipt, expectedRouterBurn *router.RouterBurn, t *testing.T) string {
-	parsedAbi, err := abi.JSON(strings.NewReader(router.RouterABI))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	routerBurn := router.RouterBurn{}
-	eventSignature := []byte("Burn(uint256,address,uint256,bytes)")
-	eventSignatureHash := crypto.Keccak256Hash(eventSignature)
-	for _, log := range txReceipt.Logs {
-		if log.Topics[0] != eventSignatureHash {
-			continue
-		}
-
-		//account := log.Topics[1]
-		//wrappedAsset := log.Topics[2]
-		err := parsedAbi.UnpackIntoInterface(&routerBurn, "Burn", log.Data)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if routerBurn.Amount.String() != expectedRouterBurn.Amount.String() {
-			t.Fatalf("Expected Burn Event Amount [%v], but actually was [%v]", expectedRouterBurn.Amount, routerBurn.Amount)
-		}
-
-		//if wrappedAsset != expectedRouterBurn.WrappedAsset.Hash() {
-		//	t.Fatalf("Expected Burn Event Wrapped Token [%v], but actually was [%v]", expectedRouterBurn.WrappedAsset, routerBurn.WrappedAsset)
-		//}
-
-		if !reflect.DeepEqual(routerBurn.Receiver, expectedRouterBurn.Receiver) {
-			t.Fatalf("Expected Burn Event Receiver [%v], but actually was [%v]", expectedRouterBurn.Receiver, routerBurn.Receiver)
-		}
-
-		//if account != expectedRouterBurn.Account.Hash() {
-		//	t.Fatalf("Expected Burn Event Account [%v], but actually was [%v]", expectedRouterBurn.Account, routerBurn.Account)
-		//}
-
-		expectedId := fmt.Sprintf("%s-%d", log.TxHash, log.Index)
-		return expectedId
-	}
-
-	t.Fatal("Could not retrieve valid Burn Event Log information.")
-	return ""
-}
-
-func validateLockEvent(txReceipt *types.Receipt, expectedRouterLock *router.RouterLock, t *testing.T) string {
-	parsedAbi, err := abi.JSON(strings.NewReader(router.RouterABI))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	routerLock := router.RouterLock{}
-	eventSignature := []byte("Lock(uint256,address,bytes,uint256,uint256)")
-	eventSignatureHash := crypto.Keccak256Hash(eventSignature)
-	for _, log := range txReceipt.Logs {
-		if log.Topics[0] != eventSignatureHash {
-			continue
-		}
-
-		err := parsedAbi.UnpackIntoInterface(&routerLock, "Lock", log.Data)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if routerLock.Amount.String() != expectedRouterLock.Amount.String() {
-			t.Fatalf("Expected Lock Event Amount [%v], actual [%v]", expectedRouterLock.Amount, routerLock.Amount)
-		}
-
-		if routerLock.TargetChain.String() != expectedRouterLock.TargetChain.String() {
-			t.Fatalf("Expected Lock Event Target Chain [%v], actual [%v]", expectedRouterLock.TargetChain, routerLock.TargetChain)
-		}
-
-		if routerLock.Token.String() != expectedRouterLock.Token.String() {
-			t.Fatalf("Expected Lock Event Token [%v], actual [%v]", expectedRouterLock.Token, routerLock.Token)
-		}
-
-		if routerLock.ServiceFee.String() != expectedRouterLock.ServiceFee.String() {
-			t.Fatalf("Expected Lock Event Service Fee [%v], actual [%s]", expectedRouterLock.ServiceFee, routerLock.ServiceFee)
-		}
-
-		if !reflect.DeepEqual(routerLock.Receiver, expectedRouterLock.Receiver) {
-			t.Fatalf("Expected Lock Event Receiver [%v], actual [%v]", expectedRouterLock.Receiver, routerLock.Receiver)
-		}
-
-		expectedId := fmt.Sprintf("%s-%d", log.TxHash, log.Index)
-		return expectedId
-	}
-
-	t.Fatal("Could not retrieve valid Burn Event Log information.")
-	return ""
-}
-
-func validateBurnERC721Event(txReceipt *types.Receipt, expectedRouterLock *router.RouterBurnERC721, t *testing.T) string {
-	parsedAbi, err := abi.JSON(strings.NewReader(router.RouterABI))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	event := router.RouterBurnERC721{}
-	eventSignature := parsedAbi.Events["BurnERC721"].ID
-	for _, log := range txReceipt.Logs {
-		if log.Topics[0] != eventSignature {
-			continue
-		}
-
-		err := parsedAbi.UnpackIntoInterface(&event, "BurnERC721", log.Data)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if event.TokenId.String() != expectedRouterLock.TokenId.String() {
-			t.Fatalf("Expected Burn-ERC721 Event TokenId [%v], actual [%v]", expectedRouterLock.TokenId, event.TokenId)
-		}
-
-		if event.TargetChain.String() != expectedRouterLock.TargetChain.String() {
-			t.Fatalf("Expected Burn-ERC721 Event Target Chain [%v], actual [%v]", expectedRouterLock.TargetChain, event.TargetChain)
-		}
-
-		if !reflect.DeepEqual(event.WrappedToken, expectedRouterLock.WrappedToken) {
-			t.Fatalf("Expected Burn-ERC721 Event Token [%v], actual [%v]", expectedRouterLock.WrappedToken, event.WrappedToken)
-		}
-
-		if !reflect.DeepEqual(event.Receiver, expectedRouterLock.Receiver) {
-			t.Fatalf("Expected Burn-ERC721 Event Receiver [%v], actual [%v]", expectedRouterLock.Receiver, event.Receiver)
-		}
-
-		expectedId := fmt.Sprintf("%s-%d", log.TxHash, log.Index)
-		return expectedId
-	}
-
-	t.Fatal("Could not retrieve valid Burn Event Log information.")
-	return ""
-}
-
-func validateSubmittedScheduledTx(setupEnv *setup.Setup, asset string, expectedTransfers []transaction.Transfer, t *testing.T) (transactionID, scheduleID string) {
-	receiverTransactionID, receiverScheduleID := validateScheduledTx(setupEnv, setupEnv.Clients.Hedera.GetOperatorAccountID(), asset, expectedTransfers, t)
-
-	membersTransactionID, membersScheduleID := validateMembersScheduledTxs(setupEnv, asset, expectedTransfers, t)
-
-	if receiverTransactionID != membersTransactionID {
-		t.Fatalf("Scheduled Transactions between members are different. Receiver [%s], Member [%s]", receiverTransactionID, membersTransactionID)
-	}
-
-	if receiverScheduleID != membersScheduleID {
-		t.Fatalf("Scheduled IDs between members are different. Receiver [%s], Member [%s]", receiverScheduleID, membersScheduleID)
-	}
-
-	return receiverTransactionID, receiverScheduleID
-}
-
-func validateScheduledMintTx(setupEnv *setup.Setup, account hedera.AccountID, asset string, expectedTransfers []transaction.Transfer, t *testing.T) (transactionID, scheduleID string) {
-	timeLeft := 180
-	for {
-		response, err := setupEnv.Clients.MirrorNode.GetAccountTokenMintTransactionsAfterTimestamp(account, now.UnixNano())
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(response.Transactions) > 1 {
-			t.Fatalf("[%s] - Found [%d] new transactions, must be 1.", account, len(response.Transactions))
-		}
-
-		txId, entityId := listenForTx(response, setupEnv.Clients.MirrorNode, expectedTransfers, asset, t)
-		if txId != "" && entityId != "" {
-			return txId, entityId
-		}
-
-		if timeLeft > 0 {
-			fmt.Printf("Could not find any scheduled transactions for account [%s]. Trying again. Time left: ~[%d] seconds\n", account, timeLeft)
-			timeLeft -= 10
-			time.Sleep(10 * time.Second)
-			continue
-		}
-		break
-	}
-
-	t.Fatalf("Could not find any scheduled transactions for account [%s]", setupEnv.Clients.Hedera.GetOperatorAccountID())
-	return "", ""
-}
-
-func validateScheduledBurnTx(setupEnv *setup.Setup, account hedera.AccountID, asset string, expectedTransfers []transaction.Transfer, t *testing.T) (transactionID, scheduleID string) {
-	timeLeft := 180
-	for {
-		response, err := setupEnv.Clients.MirrorNode.GetAccountTokenBurnTransactionsAfterTimestamp(account, now.UnixNano())
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(response.Transactions) > 1 {
-			t.Fatalf("[%s] - Found [%d] new transactions, must be 1.", account, len(response.Transactions))
-		}
-
-		txId, entityId := listenForTx(response, setupEnv.Clients.MirrorNode, expectedTransfers, asset, t)
-		if txId != "" && entityId != "" {
-			return txId, entityId
-		}
-
-		if timeLeft > 0 {
-			fmt.Printf("Could not find any scheduled transactions for account [%s]. Trying again. Time left: ~[%d] seconds\n", account, timeLeft)
-			timeLeft -= 10
-			time.Sleep(10 * time.Second)
-			continue
-		}
-		break
-	}
-
-	t.Fatalf("Could not find any scheduled transactions for account [%s]", setupEnv.Clients.Hedera.GetOperatorAccountID())
-	return "", ""
-}
-
-func validateScheduledNftTransfer(setupEnv *setup.Setup, token string, serialNum int64, t *testing.T) (transactionID, scheduleID string) {
-	sender := setupEnv.Clients.Hedera.GetOperatorAccountID()
-	receiver := setupEnv.BridgeAccount
-	timeLeft := 180
-
-	for {
-		response, err := setupEnv.Clients.MirrorNode.GetNftTransactions(token, serialNum)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		for _, nftTransfer := range response.Transactions {
-			if nftTransfer.Type == "CRYPTOTRANSFER" &&
-				nftTransfer.ReceiverAccountID == receiver.String() &&
-				nftTransfer.SenderAccountID == sender.String() {
-
-				scheduledTx, err := setupEnv.Clients.MirrorNode.GetScheduledTransaction(nftTransfer.TransactionID)
-				if err != nil {
-					t.Fatalf("Failed to retrieve scheduled transaction [%s]. Error: [%s]", nftTransfer.TransactionID, err)
-				}
-				for _, tx := range scheduledTx.Transactions {
-					if tx.Result == hedera.StatusSuccess.String() {
-						schedule, err := setupEnv.Clients.MirrorNode.GetSchedule(tx.EntityId)
-						if err != nil {
-							t.Fatalf("[%s] - Failed to get scheduled transaction for NFT [%s]. Error: [%s]", token, scheduleID, err)
-						}
-						return nftTransfer.TransactionID, schedule.ScheduleId
-					}
-				}
-			}
-		}
-
-		if timeLeft > 0 {
-			fmt.Printf("Could not find any scheduled transactions for account [%s]. Trying again. Time left: ~[%d] seconds\n", receiver, timeLeft)
-			timeLeft -= 10
-			time.Sleep(10 * time.Second)
-			continue
-		}
-		break
-	}
-
-	t.Fatalf("Could not find any scheduled transactions for account [%s]", setupEnv.Clients.Hedera.GetOperatorAccountID())
-	return "", ""
-}
-
-func validateScheduledNftAllowanceApprove(t *testing.T, setupEnv *setup.Setup, expectedTransactionID string, startTimestamp int64) (transactionID, scheduleID string) {
-	timeLeft := 180
-	receiver := setupEnv.Clients.Hedera.GetOperatorAccountID()
-
-	for {
-		scheduleCreates, err := setupEnv.Clients.MirrorNode.GetTransactionsAfterTimestamp(setupEnv.PayerAccount, startTimestamp, read_only.CryptoApproveAllowance)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		for _, scheduleCreate := range scheduleCreates {
-			scheduledTransaction, err := setupEnv.Clients.MirrorNode.GetScheduledTransaction(scheduleCreate.TransactionID)
-			if err != nil {
-				t.Fatalf("Could not get scheduled transaction for [%s]", scheduleCreate.TransactionID)
-			}
-
-			for _, tx := range scheduledTransaction.Transactions {
-				schedule, err := setupEnv.Clients.MirrorNode.GetSchedule(tx.EntityId)
-				if err != nil {
-					t.Fatalf("Could not get schedule entity for [%s]", tx.EntityId)
-				}
-
-				if schedule.Memo == expectedTransactionID {
-					return tx.TransactionID, tx.EntityId
-				}
-			}
-		}
-
-		if timeLeft > 0 {
-			fmt.Printf("Could not find any scheduled transactions for NFT Transfer for account [%s]. Trying again. Time left: ~[%d] seconds\n", receiver, timeLeft)
-			timeLeft -= 10
-			time.Sleep(10 * time.Second)
-			continue
-		}
-		break
-	}
-
-	t.Fatalf("Could not find any scheduled transactions for account [%s]", setupEnv.Clients.Hedera.GetOperatorAccountID())
-	return "", ""
-}
-
-func listenForTx(response *transaction.Response, mirrorNode *mirror_node.Client, expectedTransfers []transaction.Transfer, asset string, t *testing.T) (string, string) {
-	for _, transaction := range response.Transactions {
-		if transaction.Scheduled == true {
-			scheduleCreateTx, err := mirrorNode.GetTransaction(transaction.TransactionID)
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			for _, expectedTransfer := range expectedTransfers {
-				found := false
-				if asset == constants.Hbar {
-					for _, transfer := range transaction.Transfers {
-						if expectedTransfer == transfer {
-							found = true
-							break
-						}
-					}
-				} else {
-					for _, transfer := range transaction.TokenTransfers {
-						if expectedTransfer == transfer {
-							found = true
-							break
-						}
-					}
-				}
-
-				if !found {
-					t.Fatalf("[%s] - Expected transfer [%v] not found.", transaction.TransactionID, expectedTransfer)
-				}
-			}
-
-			for _, tx := range scheduleCreateTx.Transactions {
-				if tx.EntityId != "" {
-					return tx.TransactionID, tx.EntityId
-				}
-			}
-		}
-	}
-	return "", ""
-}
-
-func validateScheduledTx(setupEnv *setup.Setup, account hedera.AccountID, asset string, expectedTransfers []transaction.Transfer, t *testing.T) (transactionID, scheduleID string) {
-	timeLeft := 180
-	for {
-		response, err := setupEnv.Clients.MirrorNode.GetAccountCreditTransactionsAfterTimestamp(account, now.UnixNano())
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(response.Transactions) > 1 {
-			t.Fatalf("[%s] - Found [%d] new transactions, must be 1.", account, len(response.Transactions))
-		}
-
-		txId, entityId := listenForTx(response, setupEnv.Clients.MirrorNode, expectedTransfers, asset, t)
-		if txId != "" && entityId != "" {
-			return txId, entityId
-		}
-
-		if timeLeft > 0 {
-			fmt.Printf("Could not find any scheduled transactions for account [%s]. Trying again. Time left: ~[%d] seconds\n", setupEnv.Clients.Hedera.GetOperatorAccountID(), timeLeft)
-			timeLeft -= 10
-			time.Sleep(10 * time.Second)
-			continue
-		}
-		break
-	}
-
-	t.Fatalf("Could not find any scheduled transactions for account [%s]", setupEnv.Clients.Hedera.GetOperatorAccountID())
-	return "", ""
-}
-
-func validateMembersScheduledTxs(setupEnv *setup.Setup, asset string, expectedTransfers []transaction.Transfer, t *testing.T) (transactionID, scheduleID string) {
-	if len(setupEnv.Members) == 0 {
-		return "", ""
-	}
-
-	var transactions []string
-	var scheduleIDs []string
-	for _, member := range setupEnv.Members {
-		txID, scheduleID := validateScheduledTx(setupEnv, member, asset, expectedTransfers, t)
-		transactions = append(transactions, txID)
-
-		if !util.AllSame(transactions) {
-			t.Fatalf("Transaction [%s] does not match with previously added transactions.", txID)
-		}
-		scheduleIDs = append(scheduleIDs, scheduleID)
-
-		if !util.AllSame(scheduleIDs) {
-			t.Fatalf("ScheduleID [%s] does not match with previously added ids", scheduleID)
-		}
-	}
-
-	return transactions[0], scheduleIDs[0]
-}
-
-func calculateReceiverAndFeeAmounts(setup *setup.Setup, token string, amount int64) (receiverAmount, fee int64) {
-	fee, remainder := setup.Clients.FeeCalculator.CalculateFee(token, amount)
-	validFee := setup.Clients.Distributor.ValidAmount(fee)
-	if validFee != fee {
-		remainder += fee - validFee
-	}
-
-	return remainder, validFee
-}
-
-func submitMintTransaction(evm setup.EVMUtils, txId string, transactionData *service.FungibleTransferData, tokenAddress common.Address, t *testing.T) common.Hash {
-	var signatures [][]byte
-	for i := 0; i < len(transactionData.Signatures); i++ {
-		signature, err := hex.DecodeString(transactionData.Signatures[i])
-		if err != nil {
-			t.Fatalf("Failed to decode signature with error: [%s]", err)
-		}
-		signatures = append(signatures, signature)
-	}
-	mintAmount, ok := new(big.Int).SetString(transactionData.Amount, 10)
-	if !ok {
-		t.Fatalf("Could not convert mint amount [%s] to big int", transactionData.Amount)
-	}
-
-	res, err := evm.RouterContract.Mint(
-		evm.KeyTransactor,
-		new(big.Int).SetUint64(transactionData.SourceChainId),
-		[]byte(txId),
-		tokenAddress,
-		evm.Receiver,
-		mintAmount,
-		signatures,
-	)
-
-	if err != nil {
-		t.Fatalf("Cannot execute transaction - Error: [%s].", err)
-	}
-	return res.Hash()
-}
-
-func submitMintERC721Transaction(evm setup.EVMUtils, txId string, transactionData *service.NonFungibleTransferData, t *testing.T) common.Hash {
-	var signatures [][]byte
-	for i := 0; i < len(transactionData.Signatures); i++ {
-		signature, err := hex.DecodeString(transactionData.Signatures[i])
-		if err != nil {
-			t.Fatalf("Failed to decode signature with error: [%s]", err)
-		}
-		signatures = append(signatures, signature)
-	}
-
-	res, err := evm.RouterContract.MintERC721(
-		evm.KeyTransactor,
-		new(big.Int).SetUint64(transactionData.SourceChainId),
-		[]byte(txId),
-		common.HexToAddress(transactionData.TargetAsset),
-		big.NewInt(transactionData.TokenId),
-		transactionData.Metadata,
-		evm.Receiver,
-		signatures,
-	)
-
-	if err != nil {
-		t.Fatalf("Cannot execute transaction - Error: [%s].", err)
-	}
-	return res.Hash()
-}
-
-// Depricated. Used only for test Test_EVM_Wrapped_to_EVM_Token
-func submitUnlockTransaction(evm setup.EVMUtils, txId string, transactionData *service.FungibleTransferData, tokenAddress common.Address, t *testing.T) common.Hash {
-	var signatures [][]byte
-
-	for i := 0; i < len(transactionData.Signatures); i++ {
-		signature, err := hex.DecodeString(transactionData.Signatures[i])
-		if err != nil {
-			t.Fatalf("Failed to decode signature with error: [%s]", err)
-		}
-		signatures = append(signatures, signature)
-	}
-
-	mintAmount, ok := new(big.Int).SetString(transactionData.Amount, 10)
-	if !ok {
-		t.Fatalf("Could not convert mint amount [%s] to big int", transactionData.Amount)
-	}
-
-	res, err := evm.RouterContract.Unlock(
-		evm.KeyTransactor,
-		new(big.Int).SetUint64(transactionData.SourceChainId),
-		[]byte(txId),
-		tokenAddress,
-		mintAmount,
-		evm.Receiver,
-		signatures,
-	)
-
-	if err != nil {
-		t.Fatalf("Cannot execute transaction - Error: [%s].", err)
-	}
-	return res.Hash()
-}
-
-func submitUnlockWithFeeTransaction(evm setup.EVMUtils, txId string, transactionData *service.FungibleTransferData, tokenAddress common.Address, t *testing.T) common.Hash {
-	var signatures [][]byte
-
-	for i := 0; i < len(transactionData.Signatures); i++ {
-		signature, err := hex.DecodeString(transactionData.Signatures[i])
-		if err != nil {
-			t.Fatalf("Failed to decode signature with error: [%s]", err)
-		}
-		signatures = append(signatures, signature)
-	}
-
-	mintAmount, ok := new(big.Int).SetString(transactionData.Amount, 10)
-	if !ok {
-		t.Fatalf("Could not convert mint amount [%s] to big int", transactionData.Amount)
-	}
-
-	feeAmount, ok := new(big.Int).SetString(transactionData.Fee, 10)
-	if !ok {
-		t.Fatalf("Could not convert fee amount [%s] to big int", transactionData.Fee)
-	}
-
-	res, err := evm.RouterContract.UnlockWithFee(
-		evm.KeyTransactor,
-		new(big.Int).SetUint64(transactionData.SourceChainId),
-		[]byte(txId),
-		tokenAddress,
-		mintAmount,
-		evm.Receiver,
-		feeAmount,
-		signatures,
-	)
-
-	if err != nil {
-		t.Fatalf("Cannot execute transaction - Error: [%s].", err)
-	}
-	return res.Hash()
-}
-
-func generateMirrorNodeExpectedTransfersForBurnEvent(setupEnv *setup.Setup, asset string, amount, fee int64) []transaction.Transfer {
-	total := amount + fee
-	feePerMember := fee / int64(len(setupEnv.Members))
-
-	var expectedTransfers []transaction.Transfer
-	expectedTransfers = append(expectedTransfers, transaction.Transfer{
-		Account: setupEnv.BridgeAccount.String(),
-		Amount:  -total,
-	},
-		transaction.Transfer{
-			Account: setupEnv.Clients.Hedera.GetOperatorAccountID().String(),
-			Amount:  amount,
-		})
-
-	for _, member := range setupEnv.Members {
-		expectedTransfers = append(expectedTransfers, transaction.Transfer{
-			Account: member.String(),
-			Amount:  feePerMember,
-		})
-	}
-
-	if asset != constants.Hbar {
-		for i := range expectedTransfers {
-			expectedTransfers[i].Token = asset
-		}
-	}
-
-	return expectedTransfers
-}
-
-func generateMirrorNodeExpectedTransfersForLockEvent(setupEnv *setup.Setup, asset string, amount int64) []transaction.Transfer {
-	expectedTransfers := []transaction.Transfer{
-		{
-			Account: setupEnv.BridgeAccount.String(),
-			Amount:  -amount,
-			Token:   asset,
-		},
-		{
-			Account: setupEnv.Clients.Hedera.GetOperatorAccountID().String(),
-			Amount:  amount,
-			Token:   asset,
-		},
-	}
-
-	return expectedTransfers
-}
-
-func generateMirrorNodeExpectedTransfersForHederaTransfer(setupEnv *setup.Setup, asset string, fee int64) []transaction.Transfer {
-	feePerMember := fee / int64(len(setupEnv.Members))
-
-	var expectedTransfers []transaction.Transfer
-	expectedTransfers = append(expectedTransfers, transaction.Transfer{
-		Account: setupEnv.BridgeAccount.String(),
-		Amount:  -fee,
-	})
-
-	for _, member := range setupEnv.Members {
-		expectedTransfers = append(expectedTransfers, transaction.Transfer{
-			Account: member.String(),
-			Amount:  feePerMember,
-		})
-	}
-
-	if asset != constants.Hbar {
-		for i := range expectedTransfers {
-			expectedTransfers[i].Token = asset
-		}
-	}
-
-	return expectedTransfers
-}
-
-func sendBurnEthTransaction(assetsService service.Assets, evm setup.EVMUtils, asset string, sourceChainId, targetChainId uint64, receiver []byte, amount int64, t *testing.T) (*types.Receipt, *router.RouterBurn) {
-	wrappedAsset, err := setup.NativeToWrappedAsset(assetsService, sourceChainId, targetChainId, asset)
-	if err != nil {
-		t.Fatal(err)
-	}
-	fmt.Printf("Parsed [%s] to ETH Token [%s]\n", asset, wrappedAsset)
-
-	approvedValue := big.NewInt(amount)
-
-	instance, err := setup.InitAssetContract(wrappedAsset, evm.EVMClient)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	approveTx, err := instance.Approve(evm.KeyTransactor, evm.RouterAddress, approvedValue)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	fmt.Printf("[%s] Waiting for Approval Transaction\n", approveTx.Hash())
-	waitForTransaction(evm, approveTx.Hash(), t)
-
-	burnTx, err := evm.RouterContract.Burn(evm.KeyTransactor, new(big.Int).SetUint64(sourceChainId), common.HexToAddress(wrappedAsset), approvedValue, receiver)
-	if err != nil {
-		t.Fatal(err)
-	}
-	fmt.Printf("[%s] Submitted Burn Transaction\n", burnTx.Hash())
-
-	expectedRouterBurn := &router.RouterBurn{
-		//Account:      common.HexToAddress(evm.Signer.Address()),
-		//WrappedAsset: *wrappedAsset,
-		Amount:   approvedValue,
-		Receiver: receiver,
-	}
-
-	burnTxHash := burnTx.Hash()
-
-	fmt.Printf("[%s] Waiting for Burn Transaction Receipt.\n", burnTxHash)
-	burnTxReceipt, err := evm.EVMClient.WaitForTransactionReceipt(burnTxHash)
-	if err != nil {
-		t.Fatal(err)
-	}
-	fmt.Printf("[%s] Burn Transaction mined and retrieved receipt.\n", burnTxHash)
-
-	return burnTxReceipt, expectedRouterBurn
-}
-
-func sendLockEthTransaction(evm setup.EVMUtils, asset string, targetChainId uint64, receiver []byte, amount int64, t *testing.T) (*types.Receipt, *router.RouterLock) {
-	approvedValue := big.NewInt(amount)
-
-	instance, err := setup.InitAssetContract(asset, evm.EVMClient)
-	if err != nil {
-		t.Fatal(err)
-	}
-	approveTx, err := instance.Approve(evm.KeyTransactor, evm.RouterAddress, approvedValue)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	serviceFee, err := evm.RouterContract.FeeAmountFor(nil, new(big.Int).SetUint64(constants.HederaNetworkId), evm.Receiver, common.HexToAddress(asset), approvedValue)
-
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	fmt.Printf("[%s] Waiting for Approval Transaction\n", approveTx.Hash())
-	waitForTransaction(evm, approveTx.Hash(), t)
-
-	lockTx, err := evm.RouterContract.Lock(
-		evm.KeyTransactor,
-		new(big.Int).SetUint64(targetChainId),
-		common.HexToAddress(asset),
-		approvedValue,
-		receiver,
-		serviceFee,
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-	fmt.Printf("[%s] Submitted Lock Transaction\n", lockTx.Hash())
-
-	expectedRouterLock := &router.RouterLock{
-		TargetChain: new(big.Int).SetUint64(targetChainId),
-		Token:       common.HexToAddress(asset),
-		Receiver:    receiver,
-		Amount:      approvedValue,
-		ServiceFee:  serviceFee,
-		Raw:         types.Log{},
-	}
-
-	lockTxHash := lockTx.Hash()
-
-	fmt.Printf("[%s] Waiting for Lock Transaction Receipt.\n", lockTxHash)
-	lockTxReceipt, err := evm.EVMClient.WaitForTransactionReceipt(lockTxHash)
-	if err != nil {
-		t.Fatal(err)
-	}
-	fmt.Printf("[%s] Lock Transaction mined and retrieved receipt.\n", lockTxHash)
-
-	return lockTxReceipt, expectedRouterLock
-}
-
-func sendBurnERC721Transaction(evm setup.EVMUtils, wrappedToken string, targetChainId uint64, receiver []byte, serialNumber int64, t *testing.T) (*types.Receipt, *router.RouterBurnERC721) {
-	wrappedAddress := common.HexToAddress(wrappedToken)
-
-	paymentToken, err := evm.RouterContract.Erc721Payment(nil, wrappedAddress)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	fee, err := evm.RouterContract.Erc721Fee(nil, wrappedAddress)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	erc20Contract, err := wtoken.NewWtoken(paymentToken, evm.EVMClient.GetClient())
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	approveERC20Tx, err := erc20Contract.Approve(evm.KeyTransactor, evm.RouterAddress, fee)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	fmt.Printf("[%s] Waiting for ERC-20 Approval Transaction\n", approveERC20Tx.Hash())
-	waitForTransaction(evm, approveERC20Tx.Hash(), t)
-
-	erc721Contract, err := werc721.NewWerc721(wrappedAddress, evm.EVMClient)
-	if err != nil {
-		t.Fatal(err)
-	}
-	tokenId := big.NewInt(serialNumber)
-
-	approveERC721Tx, err := erc721Contract.Approve(evm.KeyTransactor, evm.RouterAddress, tokenId)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	fmt.Printf("[%s] Waiting for ERC-721 Approval Transaction\n", approveERC721Tx.Hash())
-	waitForTransaction(evm, approveERC721Tx.Hash(), t)
-	targetChainIdBigInt := new(big.Int).SetUint64(targetChainId)
-	burnTx, err := evm.RouterContract.BurnERC721(evm.KeyTransactor, targetChainIdBigInt, wrappedAddress, tokenId, paymentToken, fee, receiver)
-	if err != nil {
-		t.Fatal(err)
-	}
-	fmt.Printf("[%s] Submitted Burn Transaction\n", burnTx.Hash())
-
-	expectedRouterBurn := &router.RouterBurnERC721{
-		TargetChain:  targetChainIdBigInt,
-		WrappedToken: common.HexToAddress(wrappedToken),
-		TokenId:      tokenId,
-		Receiver:     receiver,
-	}
-
-	burnTxHash := burnTx.Hash()
-
-	fmt.Printf("[%s] Waiting for Burn ERC-721 Transaction Receipt.\n", burnTxHash)
-	burnTxReceipt, err := evm.EVMClient.WaitForTransactionReceipt(burnTxHash)
-	if err != nil {
-		t.Fatal(err)
-	}
-	fmt.Printf("[%s] Burn ERC-721 Transaction mined and retrieved receipt.\n", burnTxHash)
-
-	return burnTxReceipt, expectedRouterBurn
-}
-
-func waitForTransaction(evm setup.EVMUtils, txHash common.Hash, t *testing.T) {
-	receipt, err := evm.EVMClient.WaitForTransactionReceipt(txHash)
-	if err != nil {
-		t.Fatalf("[%s] - Error occurred while monitoring. Error: [%s]", txHash, err)
-		return
-	}
-
-	if receipt.Status == 1 {
-		fmt.Printf("TX [%s] was successfully mined\n", txHash)
-	} else {
-		t.Fatalf("TX [%s] reverted", txHash)
-	}
-}
-
-func verifyWrappedAssetBalance(evm setup.EVMUtils, nativeAsset string, mintAmount *big.Int, wrappedBalanceBefore *big.Int, wTokenReceiverAddress common.Address, t *testing.T) {
-	instance, err := setup.InitAssetContract(nativeAsset, evm.EVMClient)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	wrappedBalanceAfter, err := instance.BalanceOf(&bind.CallOpts{}, wTokenReceiverAddress)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	expectedBalance := new(big.Int).Add(wrappedBalanceBefore, mintAmount)
-
-	if wrappedBalanceAfter.Cmp(expectedBalance) != 0 {
-		t.Fatalf("Incorrect token balance. Expected to be [%s], but was [%s].", expectedBalance, wrappedBalanceAfter)
-	}
-}
-
-func verifyERC721TokenId(evm *evm.Client, wrappedToken string, serialNumber int64, receiver string, expectedMetadata string, t *testing.T) {
-	contract, err := werc721.NewWerc721(common.HexToAddress(wrappedToken), evm)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	owner, err := contract.OwnerOf(nil, big.NewInt(serialNumber))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if owner.String() != receiver {
-		t.Fatalf("Invalid owner. Expected owner for serial number [%d] to be [%s], but was [%s]", serialNumber, receiver, owner.String())
-	}
-
-	tokenURI, err := contract.TokenURI(nil, big.NewInt(serialNumber))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if expectedMetadata != tokenURI {
-		t.Fatalf("Invalid token URI. Expected token URI for serial number [%d] to be [%s], but was [%s]", serialNumber, expectedMetadata, tokenURI)
-	}
-}
-
-func validateEventTransactionIDFromValidatorAPI(setupEnv *setup.Setup, eventID, expectedTxID string, t *testing.T) {
-	actualTxID, err := setupEnv.Clients.ValidatorClient.GetEventTransactionID(eventID)
-	if err != nil {
-		t.Fatalf("[%s] - Failed to get event transaction ID. Error: [%s]", eventID, err)
-	}
-
-	if actualTxID != expectedTxID {
-		t.Fatalf("Expected Event TX ID [%s] did not match actual TX ID [%s]", expectedTxID, actualTxID)
-	}
-}
-
-func getTransactionData(setupEnv *setup.Setup, txId string, t *testing.T) []byte {
-	retryAttempts := 3
-
-	for retryAttempts > 0 {
-		bytes, err := setupEnv.Clients.ValidatorClient.GetTransferData(txId)
-		if err != nil {
-			t.Fatalf("Cannot fetch transaction data - Error: [%s].", err)
-		}
-		var transferDataResponse *service.TransferData
-		err = json.Unmarshal(bytes, &transferDataResponse)
-		if err != nil {
-			t.Fatalf("Failed to parse JSON transaction data [%s]. Error: [%s]", bytes, err)
-		}
-		if len(transferDataResponse.Signatures) == setupEnv.Scenario.ExpectedValidatorsCount {
-			return bytes
-		}
-
-		time.Sleep(10 * time.Second)
-		retryAttempts--
-	}
-
-	t.Fatalf("Failed to get transaction data for [%s] with %d signatures.", txId, setupEnv.Scenario.ExpectedValidatorsCount)
-	return nil
-}
-
-func verifyFungibleTransferFromValidatorAPI(setupEnv *setup.Setup, evm setup.EVMUtils, txId, tokenID, expectedSendAmount, targetAsset string, t *testing.T) *service.FungibleTransferData {
-	bytes := getTransactionData(setupEnv, txId, t)
-
-	var transferDataResponse *service.FungibleTransferData
-	err := json.Unmarshal(bytes, &transferDataResponse)
-	if err != nil {
-		t.Fatalf("Failed to parse JSON transaction data [%s]. Error: [%s]", bytes, err)
-	}
-
-	if transferDataResponse.IsNft {
-		t.Fatalf("Transaction data mismatch: Expected response data to not be NFT related.")
-	}
-	if transferDataResponse.Amount != expectedSendAmount {
-		t.Fatalf("Transaction data mismatch: Expected [%s], but was [%s]", expectedSendAmount, transferDataResponse.Amount)
-	}
-	if transferDataResponse.NativeAsset != tokenID {
-		t.Fatalf("Native Token mismatch: Expected [%s], but was [%s]", setupEnv.TokenID.String(), transferDataResponse.NativeAsset)
-	}
-	if transferDataResponse.Recipient != evm.Receiver.String() {
-		t.Fatalf("Receiver address mismatch: Expected [%s], but was [%s]", evm.Receiver.String(), transferDataResponse.Recipient)
-	}
-	if transferDataResponse.TargetAsset != targetAsset {
-		t.Fatalf("Token address mismatch: Expected [%s], but was [%s]", targetAsset, transferDataResponse.TargetAsset)
-	}
-
-	return transferDataResponse
-}
-
-func verifyNonFungibleTransferFromValidatorAPI(setupEnv *setup.Setup, evm setup.EVMUtils, txId, tokenID, metadata string, tokenIdOrSerialNum int64, targetAsset string, t *testing.T) *service.NonFungibleTransferData {
-	bytes := getTransactionData(setupEnv, txId, t)
-
-	var transferDataResponse *service.NonFungibleTransferData
-	err := json.Unmarshal(bytes, &transferDataResponse)
-	if err != nil {
-		t.Fatalf("Failed to parse JSON transaction data [%s]. Error: [%s]", bytes, err)
-	}
-
-	if !transferDataResponse.IsNft {
-		t.Fatalf("Transaction data mismatch: Expected response data to be NFT related.")
-	}
-	if transferDataResponse.Metadata != metadata {
-		t.Fatalf("Transaction data mismatch: Expected [%s], but was [%s]", metadata, transferDataResponse.Metadata)
-	}
-	if transferDataResponse.TokenId != tokenIdOrSerialNum {
-		t.Fatalf("Transaction tokenId/serialNum mismatch: Expected [%d], but was [%d]", tokenIdOrSerialNum, transferDataResponse.TokenId)
-	}
-	if transferDataResponse.NativeAsset != tokenID {
-		t.Fatalf("Native Token mismatch: Expected [%s], but was [%s]", setupEnv.TokenID.String(), transferDataResponse.NativeAsset)
-	}
-	if transferDataResponse.Recipient != evm.Receiver.String() {
-		t.Fatalf("Receiver address mismatch: Expected [%s], but was [%s]", evm.Receiver.String(), transferDataResponse.Recipient)
-	}
-	if transferDataResponse.TargetAsset != targetAsset {
-		t.Fatalf("Token address mismatch: Expected [%s], but was [%s]", targetAsset, transferDataResponse.TargetAsset)
-	}
-
-	return transferDataResponse
-}
-
-func verifyFeeRecord(dbValidation *database.Service, expectedRecord *entity.Fee, t *testing.T) {
-	ok, err := dbValidation.VerifyFeeRecord(expectedRecord)
-	if err != nil {
-		t.Fatalf("[%s] - Verification of database records failed - Error: [%s].", expectedRecord.TransactionID, err)
-	}
-	if !ok {
-		t.Fatalf("[%s] - Database does not contain expected fee records", expectedRecord.TransactionID)
-	}
-}
-
-func verifyTransferRecord(dbValidation *database.Service, expectedRecord *entity.Transfer, t *testing.T) {
-	exist, err := dbValidation.VerifyTransferRecord(expectedRecord)
-	if err != nil {
-		t.Fatalf("[%s] - Verification of database records failed - Error: [%s].", expectedRecord.TransactionID, err)
-	}
-	if !exist {
-		t.Fatalf("[%s] - Database does not contain expected transfer records", expectedRecord.TransactionID)
-	}
-}
-
-func verifyScheduleRecord(dbValidation *database.Service, expectedRecord *entity.Schedule, t *testing.T) {
-	exist, err := dbValidation.VerifyScheduleRecord(expectedRecord)
-	if err != nil {
-		t.Fatalf("[%s] - Verification of database records failed - Error: [%s].", expectedRecord.TransactionID, err)
-	}
-	if !exist {
-		t.Fatalf("[%s] - Database does not contain expected schedule records", expectedRecord.TransactionID)
-	}
-}
-
-func verifyTransferRecordAndSignatures(dbValidation *database.Service, expectedRecord *entity.Transfer, authMsgBytes []byte, signatures []string, t *testing.T) {
-	exist, err := dbValidation.VerifyTransferAndSignatureRecords(expectedRecord, authMsgBytes, signatures)
-	if err != nil {
-		t.Fatalf("[%s] - Verification of database records failed - Error: [%s].", expectedRecord.TransactionID, err)
-	}
-	if !exist {
-		t.Fatalf("[%s] - Database does not contain expected records", expectedRecord.TransactionID)
-	}
-}
-
-func verifyTransferToBridgeAccount(s *setup.Setup, wrappedAsset string, evm setup.EVMUtils, memo string, whbarReceiverAddress common.Address, expectedAmount int64, t *testing.T) (hedera.TransactionResponse, *big.Int) {
-	instance, err := setup.InitAssetContract(wrappedAsset, evm.EVMClient)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// Get the wrapped hbar balance of the receiver before the transfer
-	whbarBalanceBefore, err := instance.BalanceOf(&bind.CallOpts{}, whbarReceiverAddress)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	fmt.Printf("WHBAR balance before transaction: [%s]\n", whbarBalanceBefore)
-	// Get bridge account hbar balance before transfer
-	receiverBalance := util.GetHederaAccountBalance(s.Clients.Hedera, s.BridgeAccount, t).Hbars.AsTinybar()
-
-	fmt.Printf("Bridge account balance HBAR balance before transaction: [%d]\n", receiverBalance)
-
-	// Get the transaction receipt to verify the transaction was executed
-	transactionResponse, err := sendHbarsToBridgeAccount(s, memo, expectedAmount)
-	if err != nil {
-		t.Fatalf("Unable to send HBARs to Bridge Account, Error: [%s]", err)
-	}
-
-	transactionReceipt, err := transactionResponse.GetReceipt(s.Clients.Hedera)
-	if err != nil {
-		t.Fatalf("Transaction unsuccessful, Error: [%s]", err)
-	}
-
-	fmt.Printf("Successfully sent HBAR to bridge account, Status: [%s]\n", transactionReceipt.Status)
-
-	// Get bridge account hbar balance after transfer
-	receiverBalanceNew := util.GetHederaAccountBalance(s.Clients.Hedera, s.BridgeAccount, t).Hbars.AsTinybar()
-
-	fmt.Printf("Bridge Account HBAR balance after transaction: [%d]\n", receiverBalanceNew)
-
-	// Verify that the custodial address has received exactly the amount sent
-	amount := receiverBalanceNew - receiverBalance
-
-	// Verify that the bridge account has received exactly the amount sent
-	if amount != expectedAmount {
-		t.Fatalf("Expected to receive the exact transfer amount of hbar: [%v], but was [%v]", expectedAmount, amount)
-	}
-
-	return *transactionResponse, whbarBalanceBefore
-}
-
-func verifyTokenTransferToBridgeAccount(s *setup.Setup, evmAsset string, tokenID hedera.TokenID, evm setup.EVMUtils, memo string, wTokenReceiverAddress common.Address, amount int64, t *testing.T) (hedera.TransactionResponse, *big.Int) {
-	instance, err := setup.InitAssetContract(evmAsset, evm.EVMClient)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Get the wrapped hts token balance of the receiver before the transfer
-	wrappedBalanceBefore, err := instance.BalanceOf(&bind.CallOpts{}, wTokenReceiverAddress)
-	if err != nil {
-		t.Fatalf("Unable to query the token balance of the receiver account. Error: [%s]", err)
-	}
-
-	fmt.Printf("Token balance before transaction: [%s]\n", wrappedBalanceBefore)
-	// Get bridge account token balance before transfer
-	receiverBalance := util.GetHederaAccountBalance(s.Clients.Hedera, s.BridgeAccount, t)
-
-	fmt.Printf("Bridge account Token balance before transaction: [%d]\n", receiverBalance.Tokens.Get(s.TokenID))
-	// Get the transaction receipt to verify the transaction was executed
-	transactionResponse, err := sendTokensToBridgeAccount(s, tokenID, memo, amount)
-	if err != nil {
-		t.Fatalf(fmt.Sprintf("Unable to send Tokens to Bridge Account, Error: [%s]", err))
-	}
-	transactionReceipt, err := transactionResponse.GetReceipt(s.Clients.Hedera)
-	if err != nil {
-		t.Fatalf(fmt.Sprintf("Transaction unsuccessful, Error: [%s]", err))
-	}
-	fmt.Printf("Successfully sent Tokens to bridge account, Status: [%s]\n", transactionReceipt.Status)
-
-	// Get bridge account HTS token balance after transfer
-	receiverBalanceNew := util.GetHederaAccountBalance(s.Clients.Hedera, s.BridgeAccount, t)
-
-	fmt.Printf("Bridge Account Token balance after transaction: [%d]\n", receiverBalanceNew.Tokens.Get(s.TokenID))
-
-	// Verify that the custodial address has received exactly the amount sent
-	resultAmount := receiverBalanceNew.Tokens.Get(tokenID) - receiverBalance.Tokens.Get(tokenID)
-	// Verify that the bridge account has received exactly the amount sent
-	if resultAmount != uint64(amount) {
-		t.Fatalf("Expected to receive the exact transfer amount of hbar: [%v], but received: [%v]", amount, resultAmount)
-	}
-
-	return *transactionResponse, wrappedBalanceBefore
-}
-
-func verifyTopicMessages(setup *setup.Setup, txId string, t *testing.T) []string {
-	return verifyTopicMessagesWithStartTime(setup, txId, time.Now().UnixNano(), t)
-}
-
-func verifyTopicMessagesWithStartTime(setup *setup.Setup, txId string, startTime int64, t *testing.T) []string {
-	ethSignaturesCollected := 0
-	var receivedSignatures []string
-	signatureChannel := make(chan string)
-
-	fmt.Printf("Waiting for Signatures & TX Hash to be published to Topic [%v]\n", setup.TopicID.String())
-
-	// Subscribe to Topic
-	subscription, err := hedera.NewTopicMessageQuery().
-		SetStartTime(time.Unix(0, startTime)).
-		SetTopicID(setup.TopicID).
-		Subscribe(
-			setup.Clients.Hedera,
-			func(response hedera.TopicMessage) {
-				msg := &validatorproto.TopicMessage{}
-				err := proto.Unmarshal(response.Contents, msg)
-				if err != nil {
-					t.Fatal(err)
-				}
-
-				var transferID string
-				var signature string
-				switch msg.Message.(type) {
-				case *validatorproto.TopicMessage_FungibleSignatureMessage:
-					message := msg.GetFungibleSignatureMessage()
-					transferID = message.TransferID
-					signature = message.Signature
-					break
-				case *validatorproto.TopicMessage_NftSignatureMessage:
-					message := msg.GetNftSignatureMessage()
-					transferID = message.TransferID
-					signature = message.Signature
-				}
-
-				//Verify that all the submitted messages have signed the same transaction
-				if transferID != txId {
-					fmt.Printf("Expected signature message to contain the transaction id: [%s]\n", txId)
-				} else {
-					signatureChannel <- signature
-					fmt.Printf("Received Auth Signature [%s]\n", signature)
-				}
-			},
-		)
-	if err != nil {
-		t.Fatalf("Unable to subscribe to Topic [%s]", setup.TopicID)
-	}
-
-	timeoutTimer := time.NewTimer(120 * time.Second)
-
-	expectedValidatorsCount := setup.Scenario.ExpectedValidatorsCount
-
-signatureLoop:
-	for ethSignaturesCollected < expectedValidatorsCount {
-		select {
-		case signature := <-signatureChannel:
-			receivedSignatures = append(receivedSignatures, signature)
-			ethSignaturesCollected++
-		case <-timeoutTimer.C:
-			break signatureLoop
-		}
-	}
-
-	if ethSignaturesCollected != expectedValidatorsCount {
-		t.Fatalf("Expected the count of collected signatures to equal the number of validators: [%v], but was: [%v]", expectedValidatorsCount, ethSignaturesCollected)
-	}
-	subscription.Unsubscribe()
-	return receivedSignatures
-}
-
-func sendHbarsToBridgeAccount(setup *setup.Setup, memo string, amount int64) (*hedera.TransactionResponse, error) {
-	hbarSendAmount := hedera.HbarFromTinybar(amount)
-	hbarRemovalAmount := hedera.HbarFromTinybar(-amount)
-	fmt.Printf("Sending [%v] Hbars through the Bridge. Transaction Memo: [%s]\n", hbarSendAmount, memo)
-
-	res, err := hedera.NewTransferTransaction().
-		AddHbarTransfer(setup.Clients.Hedera.GetOperatorAccountID(), hbarRemovalAmount).
-		AddHbarTransfer(setup.BridgeAccount, hbarSendAmount).
-		SetTransactionMemo(memo).
-		Execute(setup.Clients.Hedera)
-	if err != nil {
-		return nil, err
-	}
-	rec, err := res.GetReceipt(setup.Clients.Hedera)
-	if err != nil {
-		return nil, err
-	}
-
-	fmt.Printf("TX broadcasted. ID [%s], Status: [%s]\n", res.TransactionID, rec.Status)
-	time.Sleep(1 * time.Second)
-
-	return &res, err
-}
-
-func sendNFTAllowance(setup *setup.Setup, nftId hedera.NftID, ownerAccountId, spenderAccountId hedera.AccountID) (*hedera.TransactionResponse, error) {
-	fmt.Printf("Sending Allowance for NFT [%s] to account [%s]\n", nftId.String(), spenderAccountId.String())
-	res, err := hedera.NewAccountAllowanceApproveTransaction().
-		ApproveTokenNftAllowance(
-			nftId,
-			ownerAccountId,
-			spenderAccountId,
-		).Execute(setup.Clients.Hedera)
-
-	if err != nil {
-		return nil, err
-	}
-	rec, err := res.GetReceipt(setup.Clients.Hedera)
-	if err != nil {
-		return nil, err
-	}
-	fmt.Printf("TX broadcasted. ID [%s], Status: [%s]\n", res.TransactionID, rec.Status)
-	time.Sleep(4 * time.Second)
-
-	return &res, err
-}
-
-func sendTokensToBridgeAccount(setup *setup.Setup, tokenID hedera.TokenID, memo string, amount int64) (*hedera.TransactionResponse, error) {
-	fmt.Printf("Sending [%v] Tokens to the Bridge. Transaction Memo: [%s]\n", amount, memo)
-
-	res, err := hedera.NewTransferTransaction().
-		SetTransactionMemo(memo).
-		AddTokenTransfer(tokenID, setup.Clients.Hedera.GetOperatorAccountID(), -amount).
-		AddTokenTransfer(tokenID, setup.BridgeAccount, amount).
-		Execute(setup.Clients.Hedera)
-	if err != nil {
-		return nil, err
-	}
-	rec, err := res.GetReceipt(setup.Clients.Hedera)
-	if err != nil {
-		return nil, err
-	}
-
-	fmt.Printf("TX broadcasted. ID [%s], Status: [%s]\n", res.TransactionID, rec.Status)
-	time.Sleep(1 * time.Second)
-
-	return &res, err
-}
-
-func sendFeeForNFTToBridgeAccount(setup *setup.Setup, receiver common.Address, chainId uint64, nftID hedera.NftID, fee int64) (*hedera.TransactionResponse, error) {
-	hbarSendAmount := hedera.HbarFromTinybar(fee)
-	hbarRemovalAmount := hedera.HbarFromTinybar(-fee)
-
-	memo := fmt.Sprintf("%d-%s-%s", chainId, receiver, nftID.String())
-	fmt.Printf("Sending Fungible Fee for NFT [%s] through the Portal. Transaction Memo: [%s]\n", nftID.String(), memo)
-
-	res, err := hedera.NewTransferTransaction().
-		AddHbarTransfer(setup.Clients.Hedera.GetOperatorAccountID(), hbarRemovalAmount).
-		AddHbarTransfer(setup.BridgeAccount, hbarSendAmount).
-		SetTransactionMemo(memo).
-		Execute(setup.Clients.Hedera)
-	if err != nil {
-		return nil, err
-	}
-	rec, err := res.GetReceipt(setup.Clients.Hedera)
-	if err != nil {
-		return nil, err
-	}
-
-	fmt.Printf("TX broadcasted. ID [%s], Status: [%s]\n", res.TransactionID, rec.Status)
-	time.Sleep(1 * time.Second)
-
-	return &res, err
 }
