+++ e2e/helper/fetch/hedera.go
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2022 LimeChain Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fetch
+
+import (
+	"testing"
+
+	"github.com/hashgraph/hedera-sdk-go/v2"
+)
+
+func HederaAccountBalance(t *testing.T, client *hedera.Client, account hedera.AccountID) hedera.AccountBalance {
+	t.Helper()
+	// Get bridge account hbar balance before transfer
+	receiverBalance, err := hedera.NewAccountBalanceQuery().
+		SetAccountID(account).
+		Execute(client)
+	if err != nil {
+		t.Fatalf("Unable to query the balance of the account [%s], Error: [%s]", account.String(), err)
+	}
+	return receiverBalance
+}
